% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_point}
\alias{as_point}
\alias{as_points}
\title{Convert a simple feature object or a numeric string to a single POINT (sfg)
or POINT geometry set (sfc)}
\usage{
as_points(...)

as_point(...)
}
\arguments{
\item{...}{See details.}
}
\description{
Works with sf, sfc, and bbox objects using \link[sf:geos_unary]{sf::st_centroid}. Works with
\link{sf_bbox_point}
}
\details{
For \link{as_point}, ... is passed to \link[sf:geos_unary]{sf::st_centroid} if ... is a sf, sfc,
or bbox object, \link{sf_bbox_point} includes a bbox object and a string
indicating the requested point position, or \link[sf:st]{sf::st_point} if ... includes a
numeric vector.

For \link{as_points} parameters are passed to as_point using \link[purrr:map]{purrr::map} and then
converted to sfc using \link[sf:st_as_sfc]{sf::st_as_sfc}. The ... parameters must include a
crs, otherwise the crs will be NA for the resulting sfc object.
}
