% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_bbox}
\title{Convert an object to a simple feature or bounding box object}
\usage{
as_sf(x, crs = NULL, ...)

as_bbox(x, crs = NULL, ...)
}
\arguments{
\item{x}{A \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, \code{sp}, or \code{dataframe} object that
can be converted into a simple feature or bounding box object. \code{\link[=as_bbox]{as_bbox()}}
can also convert a vector with xmin, ymin, xmax, and ymax values.}

\item{crs}{Coordinate reference system for sf or bbox obect to return}

\item{...}{Additional parameters passed to \code{\link[sf:st_bbox]{sf::st_bbox()}} when calling
\code{\link[=as_bbox]{as_bbox()}} or passed to \code{\link[sf:sf]{sf::st_sf()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}, or \code{\link[=df_to_sf]{df_to_sf()}}
for \code{\link[=as_sf]{as_sf()}} (depending on class of x)}
}
\description{
Both functions will pass a NULL value without returning an error.
}
