% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_data_list.R
\name{as_sf_list}
\alias{as_sf_list}
\title{Make a simple feature list with an optional grouping column}
\usage{
as_sf_list(x, nm = "data", col = NULL)
}
\arguments{
\item{x}{an sf object or a data frame with a sf list column named "data"}

\item{nm}{name(s) for sf list; defaults to "data". If col is provided, the
values of the grouping column are used as names.}

\item{col}{Grouping column must be a named varialable}
}
\description{
Make a simple feature list with an optional grouping column
}
