% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{cache_sf}
\alias{cache_sf}
\title{Cache simple feature data}
\usage{
cache_sf(
  data = NULL,
  data_dir = NULL,
  overwrite = FALSE,
  name = NULL,
  label = NULL,
  filetype = NULL,
  filename = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{data}{Data to cache.}

\item{data_dir}{cache data directory, defaults to
\code{\link[rappdirs]{user_cache_dir}} when data_dir is NULL}

\item{overwrite}{Logical. Default FALSE. If TRUE, overwrite any existing
cached files that use the same file name.}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor]{make_clean_case}}, e.g. "Residential zoning map"
becomes "residential_zoning_map"}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor]{make_clean_case}}. The label is designed to identify
the area or other shared characteristics across multiple data files, maps,
or plots.}

\item{filetype}{File type or extension.}

\item{filename}{File name to use for cached file. Defaults to name of data.
If the data is an sf object make sure to include the file type, e.g.
"data.gpkg", supported by \code{sf::write_sf()}. All other data is written to
rda with \code{readr::write_rds()}.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}
}
\description{
Cache data to \code{rappdirs::user_cache_dir("overedge")} or other data directory.
}
