% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{convert_dist_scale}
\alias{convert_dist_scale}
\title{Convert distance from scale to actual units}
\usage{
convert_dist_scale(
  dist = NULL,
  paper = NULL,
  orientation = NULL,
  scale = NULL,
  scale_unit = "in",
  scale_factor = NULL,
  actual_unit = NULL,
  dpi = 120
)
}
\arguments{
\item{dist}{distance to convert. If paper is provided, paper width and height
are used as dist.}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{scale}{Scale name}

\item{scale_unit}{"mm" (converted to cm by dividing by 10), "cm", "px"
(converted to inches by dividing by dpi), or "in".}

\item{actual_unit}{any unit supported by convert_dist_units}

\item{dpi}{dots per square inch (used as conversion factor for "px" to "in")}
}
\value{
dist values converted from scale_unit to actual_unit based on
scale_factor or information from standard_scales object. If paper is
provided, return a data frame with converted distances as actual_width and
actual_height
}
\description{
Convert distance from scale to actual units based on named \code{standard_scales}
}
