% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_count.R
\name{count_features}
\alias{count_features}
\title{Count simple feature objects}
\usage{
count_features(
  data = NULL,
  boundaries = NULL,
  join = NULL,
  by = NULL,
  .id = "count",
  ...
)
}
\arguments{
\item{data}{Data frame or \code{sf} object, Default: NULL}

\item{boundaries}{Boundary data as an \code{sf} object, Default: NULL}

\item{by}{Character string to join data by if data is not an \code{sf} object, Default: NULL}

\item{.id}{Count column, Default: "count"}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Count the number of simple feature objects within a geometry or by some
attribute contained in both the data and the corresponding boundary geometry.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{geos_binary_pred}}, \code{\link[sf]{st_nearest_feature}}, \code{\link[sf]{st_geometry}}, \code{\link[sf]{st_join}}, \code{\link[sf]{st_as_sf}}
\code{\link[purrr]{map}}
\code{\link[dplyr]{pull}}, \code{\link[dplyr]{mutate-joins}}
}
