% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_sf.R
\name{df_to_sf}
\alias{df_to_sf}
\title{Convert data frame with to sf object}
\usage{
df_to_sf(
  x,
  coords = c("LONGITUD", "LATITUDE"),
  into = NULL,
  sep = ",",
  lonlat = TRUE,
  crs = 4326
)
}
\arguments{
\item{x}{data frame with lat/lon coordinates in 1 combined or 2 separated columns}

\item{coords}{character string with names of longitude and latitude column or columns, Default: c('LONGITUD', 'LATITUDE')}

\item{into}{If coords is a single column name with both longitude and latitude, \code{into} is used as the names of the new columns that coords is separated into. Passed to tidyr::separate().}

\item{sep}{If coords is a single column name with both longitude and latitude, \code{sep} is used as the separator between coordinate values. Passed to tidyr::separate().}

\item{lonlat}{If coords are ordered as latitude, longitude order or if into is latitude, longitude order, set to FALSE. Default TRUE.}

\item{crs}{coordinate reference system, Default: 4326}
}
\value{
sf object
}
\description{
Convert data frame with to sf object
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
  nc_df <- ggspatial::df_spatial(nc)

  df_as_sf(nc_df, coords = c("x", "y"))

  nc_df$xy <- paste(nc_df$x, nc_df$y, sep = ",")

  df_as_sf(nc_df, coords = "xy", into = c("lon", "lat"))
}
}
# TODO: reference this one? https://docs.ropensci.org/stats19/reference/format_sf.html
}
\seealso{
\code{\link[ggspatial]{df_spatial}},
\code{\link[sf]{st_as_sf}}
\code{\link[stats19]{format_sf}}
}
