% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\alias{fix_date}
\alias{relocate_sf_col}
\alias{rename_sf_col}
\alias{bind_address_col}
\alias{bind_units_col}
\title{Format data using common approaches}
\usage{
format_data(
  x,
  var_names = NULL,
  clean_names = TRUE,
  replace_na = NULL,
  replace_with_na = NULL,
  replace_empty_char_with_na = TRUE,
  fix_date = TRUE
)

fix_date(x)

relocate_sf_col(x, .after = dplyr::everything())

rename_sf_col(x, sf_col = "geometry")

bind_address_col(x, city = NULL, county = NULL, state = NULL)

bind_units_col(x, y, units = NULL, drop = FALSE, keep_all = TRUE, .id = NULL)
}
\arguments{
\item{x}{A tibble or data frame object}

\item{var_names}{A named list following the format, list("New var name" =
old_var_name), or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}

\item{clean_names}{Optionally apply \link[janitor:clean_names]{janitor::clean_names}; defaukts to \code{TRUE}.}

\item{replace_na}{A named list to pass to \link[tidyr:replace_na]{tidyr::replace_na}; defaults to
\code{NULL}.}

\item{replace_with_na}{A named list to pass to \link[naniar:replace_with_na]{naniar::replace_with_na};
defaults to \code{NULL}.}

\item{replace_empty_char_with_na}{If \code{TRUE}, replace "" with \code{NA} using
\link[naniar:replace_with_na_if]{naniar::replace_with_na_if}, Default: \code{TRUE}}

\item{fix_date}{If \code{TRUE}, fix UNIX dates (common issue with dates from
FeatureServer and MapServer sources) , Default: \code{TRUE}}

\item{.after}{The location to place sf column after; defaults to
\link[dplyr:reexports]{dplyr::everything}.}

\item{sf_col}{Name to rename sf column to.}

\item{city, county, state}{City, county, and state to bind to data frame or
\code{sf} object.}

\item{keep_all}{If \code{FALSE}, keep all columns. If FALSE, return only the named
.id column.}
}
\value{
Data with formatting functions applied.
}
\description{
This function can apply the following common data cleaning tasks:
}
\details{
\itemize{
\item Applies \link[stringr:str_trim]{stringr::str_squish} and \link[stringr:str_trim]{stringr::str_trim} to all character columns
\item Optionally replaces all character values of "" with NA values
\item Optionally corrects UNIX formatted dates with 1970-01-01 origins
\item Optionally renames variables by passing a named list of variables
}
}
