% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_icon.R
\name{geom_sf_icon}
\alias{geom_sf_icon}
\title{Draw SVG icons for sf objects}
\usage{
geom_sf_icon(
  data = NULL,
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  ...
)
}
\arguments{
\item{data}{sf object. If the geometry type for data is POINT, the object is
used as is. If not, the object is converted to POINT using
sf::st_centroid(). If the sf object has a column named "icon" and \code{icon} is
NULL, the value of the column will be used as the icon name.}

\item{icon}{Icon name. If the data includes a column named icon, \code{icon} is
optional. Otherwise, \code{icon} is required.}

\item{px}{Icon size in pixels. Optional but may be required to differentiate
icons with duplicate names.}

\item{source}{Icon source. Options include "mapbox/maki", "ideditor/temaki",
"manifestinteractive/weather-underground-icons", or
"Esri/calcite-point-symbols". Optional but may be required to differentiate
icons with duplicate names.}

\item{...}{Parameters to ggsvg::geom_point_svg()}
}
\description{
Use the \code{ggsvg::geom_point_svg()} function to plot icons at
coordinates for an sf object.
}
\examples{
\dontrun{
if(interactive()){
 library(sf)
 library(ggplot2)
 library(overedge)

 nc <- st_read(system.file("shape/nc.shp", package="sf"))
 ggplot() +
  geom_sf_icon(data = nc, icon = "point-start", size = 10)


 nc$icon <- rep(c("1", "2", "3", "4"), nrow(nc)/4)
 ggplot() +
   geom_sf_icon(data = nc, size = 5)
 }
}
}
\seealso{
\code{\link[ggsvg]{geom_point_svg}}
\code{\link[overedge]{map_icons}}
}
