% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_asp}
\alias{get_asp}
\title{Get aspect ratio from string or based on specific paper and margins}
\usage{
get_asp(
  asp = NULL,
  paper = NULL,
  orientation = NULL,
  margin = NULL,
  unit = NULL,
  block_asp = FALSE,
  null.ok = TRUE
)
}
\arguments{
\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{margin}{A margin style supported by \code{\link[=get_margin]{get_margin()}}, a numeric vector
(length 1 or 4) passed to dist parameter of get_margin, or a margins
object.}

\item{unit}{Unit for margin distance, Default: 'in'.}

\item{block_asp}{If \code{TRUE}, get aspect ratio of the map/plot area (not
including the margins); defaults to \code{FALSE}.}
}
\value{
A numeric aspect ratio.
}
\description{
Get aspect ratio from string or based on specific paper and margins
}
\examples{
get_asp("1:2")

get_asp(11 / 17)

get_asp(paper = "letter")
}
