% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Check existence of data directory and create if nonexistent}
\usage{
get_data_dir(path = NULL)
}
\arguments{
\item{path}{Path to directory to use as data directory.}
}
\description{
Check existence of data directory and create if nonexistent
}
