% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_data.R
\name{get_esri_data}
\alias{get_esri_data}
\title{Get data from an ArcGIS FeatureServer or MapServer}
\usage{
get_esri_data(
  location = NULL,
  url,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = NULL,
  where = NULL,
  coords_col = NULL,
  lonlat = TRUE,
  locationname_col = NULL,
  locationname,
  ...
)
}
\arguments{
\item{location}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \code{esri2sf::esri2sf()} function.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{coordinate reference system to use for query; default 4326}

\item{where}{PARAM_DESCRIPTION, Default: '1=1'}

\item{lonlat}{Default TRUE. If FALSE, coords is assumed to be lat/lon order instead of lon/lat.}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}
}
\description{
Wraps the \code{esri2sf::esri2sf()} function to download an ArcGIS
FeatureServer or MapServer.
}
\seealso{
\code{\link[esri2sf]{esri2sf}}
}
