% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_data.R
\name{get_osm_data}
\alias{get_osm_data}
\title{Use osmdata to get Open Street Map data for a location}
\usage{
get_osm_data(
  location = NULL,
  key,
  value = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = NULL,
  geometry = NULL,
  osmdata = FALSE
)
}
\arguments{
\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{key}{Feature key for overpass API query.}

\item{value}{Value of the feature key; can be negated with an initial
exclamation mark, \code{value = "!this"}, and can also be a vector, \code{value = c("this", "that")}.
If \code{value = "all"} or if \code{key = "building"} the values passed to the
osmdata package are from
\code{\link[osmdata]{available_tags()}}.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Units are passed to
\code{units::set_units()} and then converted to units matching GDAL; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system for output data; if NULL, the data
remains in the Open Street Map coordinate reference system 4326. Default:
\code{NULL}.}

\item{geometry}{Geometry type to output ("polygons", "points", "lines",
"multilines", or "multipolygons"); if multiple geometry types are needed
set osmdata to \code{TRUE.} Default \code{NULL}.}

\item{osmdata}{If \code{TRUE} return a \code{osmdata} class object that includes the
overpass API call, metadata including timestamp and version numbers, and
all available geometry types; defaults to \code{FALSE}.}
}
\value{
A simple feature object with features using selected geometry type or
an \code{osmdata} object with features from all geometry types.
}
\description{
Wraps \code{osmdata} functions.
}
