% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_socrata_data.R
\name{get_socrata_data}
\alias{get_socrata_data}
\title{Use RSocrata to get data from a Socrata open data portal for a location}
\usage{
get_socrata_data(
  resource = NULL,
  select = NULL,
  where = NULL,
  query = NULL,
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  locationname_col = NULL,
  locationname = NULL,
  coords = c("longitude", "latitude"),
  geometry = FALSE,
  site_url = NULL,
  key = NULL,
  from_crs = 4326,
  crs = NULL
)
}
\arguments{
\item{resource}{A Socrata dataset identifier. If resource is set to "list"
and a valid site_url is provided, the function returns a list of all
available resources.}

\item{select}{SODA $select parameter. Set of columns to be returned, similar
to a SELECT in SQL. \url{https://dev.socrata.com/docs/queries/select.html}}

\item{where}{SODA $where parameter. Filters the rows to be returned, similar
to WHERE. \url{https://dev.socrata.com/docs/queries/where.html}}

\item{query}{SODA $query parameter. A full SoQL query string, all as one
parameter. \url{https://dev.socrata.com/docs/queries/query.html}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Units are passed to
\code{units::set_units()} and then converted to units matching GDAL; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{locationname_col}{name of column in Socrata resource with
location names (e.g. County)}

\item{locationname}{location name to return}

\item{coords}{Name(s) of column with coordinate data, Default: c("longitude",
"latitude")}

\item{geometry}{If \code{TRUE} and latitude/longitude columns available, return a
\code{\link{sf}} object. Default \code{FALSE}.}

\item{site_url}{A Socrata site url.}

\item{crs}{Coordinate reference system to return.}

\item{trim}{If area is provided, trim data to the area boundary rather than
the bounding box, Default: FALSE. area must be provided if TRUE.}
}
\description{
Get a selected dataset using Socrata Query Language (SoQL) parameters as a
tibble or sf object. Details on SoQL queries are found in the Socrata API
documentation \url{https://dev.socrata.com/docs/queries/}
}
\examples{
\dontrun{
## Get Q2 2020 vehicle crash data for Cecil County, Maryland
get_socrata_data(
  site_url = "https://opendata.maryland.gov",
  resource = "65du-s3qu",
  where = "(year = '2020') AND (quarter = 'Q2')",
  locationname_col = "county_desc",
  locationname = "Cecil"
)
}
}
