% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_standard_scale}
\alias{get_standard_scale}
\title{Get standard scales and convert to scale distances}
\usage{
get_standard_scale(
  standard = NULL,
  series = NULL,
  scale = NULL,
  paper = NULL,
  orientation = NULL,
  convert = FALSE
)
}
\arguments{
\item{standard}{USGS, Engineering, or Architectural}

\item{series}{Map series}

\item{scale}{Scale name}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{convert}{If \code{TRUE}, return paper with scale distances converted to actual distances based on values from selected scale; defaults to \code{FALSE}.}
}
\description{
Get standard scales and convert to scale distances
}
