% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_boundaries.R
\name{get_us_counties}
\alias{get_us_counties}
\title{Use USAboundariesData to get current or past U.S. county boundaries for a location}
\usage{
get_us_counties(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  historical = FALSE,
  crs = NULL,
  crop = FALSE,
  trim = FALSE,
  detailed = FALSE
)
}
\arguments{
\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{historical}{If \code{TRUE}, return historical data (use
\code{USAboundaries::us_counties()} if returning geometry for a specific
map_date is required).}

\item{crs}{coordinate reference system to return}

\item{crop}{If TRUE, data cropped to location or bounding box
\code{\link[sf]{st_crop}} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp}
parameters provided. Default TRUE.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default FALSE.}

\item{detailed}{If \code{TRUE}, download "hires" version of county boundaries. (Not
currently supported).}
}
\value{
A \code{sf} object with county boundaries (contemporary or historical)
}
\description{
Wraps {USAboundariesData} us counties data to effectively make a version of
the \code{USAboundaries::us_counties()} that works with an adjusted location as a
spatial filter but does not support the support the map_date parameter for
historical county boundary data.
}
\seealso{
\code{\link[tigris]{states}}
\code{\link[USAboundaries]{us_counties}}
}
