% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wiki_data.R
\name{get_wiki_data}
\alias{get_wiki_data}
\title{Get Wikipedia articles for a location}
\usage{
get_wiki_data(
  location,
  dist = 100,
  diag_ratio = NULL,
  asp = NULL,
  radius = FALSE,
  primary = TRUE,
  details = c("type", "name", "region"),
  lang = "en",
  limit = 50,
  geometry = TRUE
)
}
\arguments{
\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{radius}{If TRUE, use dist as a buffer around the center of the
location.}

\item{primary}{If TRUE, search for primary coordinates. Set primary to "all"
or "secondary" to search other coordinate types.}

\item{details}{Additional detailed to return with results. Options include
"type", "name", "country", "region"}

\item{lang}{Language to search on Wikipedia; defaults to "en".}

\item{limit}{Number of pages to return (max 500); deafaults to 50}

\item{geometry}{If TRUE, return sf object. If FALSE, return dataframe. Defaults to FALSE.}
}
\description{
Use the Wikipedia API geosearch API to get Wikipedia articles for a location.
See \url{https://www.mediawiki.org/wiki/Extension:GeoData} for more information.
Only returns Wikipedia articles with coordinates.
}
