% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_ext.R
\name{ggsave_ext}
\alias{ggsave_ext}
\title{Save a ggplot and update file EXIF metadata}
\usage{
ggsave_ext(
  name = NULL,
  label,
  plot = last_plot(),
  filename = NULL,
  path = NULL,
  device,
  scale = 1,
  asp,
  width,
  height,
  units = "in",
  dpi,
  exif = TRUE,
  title = NULL,
  author,
  args = NULL,
  ...
)
}
\arguments{
\item{name}{name of plot, used to create filename}

\item{label}{label is appending to filename as a prefix.}

\item{asp}{aspect ratio}

\item{exif}{If TRUE, edit exif metadata for exported file, Default: TRUE}

\item{title}{title of plot or map, added to exif metadata, Default: NULL}

\item{author}{author of plot or map, added to exif metadata,}

\item{args}{args passed to exiftoolr, If args is not NULL, title and author
are ignored, Default: NULL}

\item{...}{additional parameters passed to \code{\link[ggplot2]{ggsave}}}
}
\description{
Save a plot or map then update the EXIF metadata for the title, author, and
create data. \code{ggsave_ext()} also supports creating a file name based on a
sentence case name with spaces (e.g. "Baltimore city map") and appending a
label (e.g. "baltcity") as a prefix to the output file name.
}
\seealso{
\code{\link[exifr]{exiftool_call}}
}
