% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_same_name_col}
\alias{has_same_name_col}
\title{Does the data frame has a column with the same name?}
\usage{
has_same_name_col(
  x,
  col = NULL,
  prefix = "orig",
  ask = FALSE,
  quiet = FALSE,
  drop = TRUE
)
}
\description{
Does the data frame has a column with the same name?
}
