% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_units.R
\name{is_dist_units}
\alias{is_dist_units}
\alias{get_dist_units}
\alias{as_dist_units}
\alias{diff_dist}
\alias{is_diff_dist}
\alias{is_same_dist}
\alias{is_longer}
\alias{is_shorter}
\title{General utility functions for working with distance units objects}
\usage{
is_dist_units(x)

get_dist_units(x, null.ok = TRUE)

as_dist_units(x, units = NULL, null.ok = FALSE)

is_diff_dist(x, y, units = NULL)

is_same_dist(x, y, dist = NULL, diff = FALSE, ...)

is_longer(x, y)

is_shorter(x, y)
}
\arguments{
\item{x}{object to check}
}
\description{
\itemize{
\item is_dist_units: Is x a distance unit object?
\item get_dist_units: Get distance units from x (works for sf and units objects)
\item is_same_dist, is_shorter, is_longer: Compare units objects
}
}
