% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_geom_type.R
\name{is_geom_type}
\alias{is_geom_type}
\alias{is_point}
\alias{is_multipoint}
\alias{is_string}
\alias{is_multistring}
\alias{is_polygon}
\alias{is_multipolygon}
\title{What geometry type is this feature?}
\usage{
is_geom_type(x, ext = TRUE, type = NULL, by_geometry = FALSE)

is_point(x, by_geometry = FALSE)

is_multipoint(x, by_geometry = FALSE)

is_string(x, by_geometry = FALSE)

is_multistring(x, by_geometry = FALSE)

is_polygon(x, by_geometry = FALSE)

is_multipolygon(x, by_geometry = FALSE)
}
\arguments{
\item{x}{A sf or sfc object passed to \link[sf:st_geometry_type]{sf::st_geometry_type}}

\item{ext}{For st_geom_type, if ext TRUE and check is NULL, return a list with checks for POINTS,
POLYGONS, LINESTRING, and the returned types.}

\item{type}{If "POINT", check if geometry type is POINT. Same for all
available geometry types; not case sensitive; Default: NULL}

\item{by_geometry}{Passed to sf::st_geometry_type; defaults to FALSE}
}
\value{
Returns vector with all geometry types; gives warning if object uses
multiple types.
}
\description{
A flexible wrapper for \link[sf:st_geometry_type]{sf::st_geometry_type}.
}
