% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gg.R
\name{is_gg}
\alias{is_gg}
\alias{is_gg_list}
\alias{is_gg_layer}
\alias{is_gg_sf_layer}
\alias{is_gg_scale}
\alias{is_gg_theme}
\alias{is_gg_component}
\title{Check ggplot objects and ggplot lists}
\usage{
is_gg_list(x)

is_gg_layer(x)

is_gg_sf_layer(x)

is_gg_scale(x)

is_gg_theme(x)

is_gg_component(x)
}
\description{
Check ggplot objects and ggplot lists
}
