% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_sf}
\alias{is_sf}
\alias{is_sfc}
\alias{is_bbox}
\alias{is_sf_list}
\alias{is_same_crs}
\alias{is_raster}
\alias{is_sp}
\title{Check the class and attributes of a simple feature object}
\usage{
is_sf(x, ext = FALSE)

is_sfc(x)

is_bbox(x)

is_sf_list(x, is_named = FALSE)

is_same_crs(x, y)

is_raster(x)

is_sp(x)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by \link{is_sf})}

\item{is_named}{If TRUE, check if sf list is named; defaults FALSE}

\item{y}{An sf object or a character or numeric object supported by
\link[sf:st_crs]{sf::st_crs} that can be compared to x. (used by \link{is_same_crs})}
}
\description{
Check the class and attributes of a simple feature object
}
\details{
\itemize{
\item \link{is_sf}: check if x it a \code{sf} class object.
\item \link{is_bbox}: check if x is a \code{bbox} class object.
\item \link{is_sf_list}: check if x is a named list of \code{sf} class objects.
}
}
