% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_data.R
\name{layer_location_data}
\alias{layer_location_data}
\title{Layer location data into a ggplot2 map}
\usage{
layer_location_data(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  package = NULL,
  filetype = NULL,
  fn = NULL,
  geom_fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = NULL,
  crs = NULL,
  label_col = "name",
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\link[ggplot2:ggsf]{ggplot2::geom_sf}), "icon" (\link{layer_icon}),
"markers" (\link{layer_markers}), "text" (\link[ggplot2:ggsf]{ggplot2::geom_sf_text}), and "label"
(\link[ggplot2:ggsf]{ggplot2::geom_sf_label}). See details for a full list.}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \link[sf:geos_combine]{sf::st_union}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{unit to adjust location by dist or diag_ratio; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{package}{package name.}

\item{filetype}{file type supported by \link[sf:st_read]{sf::read_sf}. The file type must be
provided for extdata and cached data.}

\item{fn}{Function to apply to data before returning.}

\item{geom_fn}{ggplot2 geom or custom function using lambda syntax. Use for
passing custom mapping functions to layer_location_data beyond the
supported geom options.}

\item{crop}{If \code{TRUE}, data is cropped to location or bounding box
\link[sf:st_crop]{sf::st_crop} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp} parameters
provided. Default TRUE.}

\item{trim}{If \code{TRUE}, data is trimmed to area with \link[sf:geos_binary_ops]{sf::st_intersection}.
This option ignores any \code{dist}, \code{diag_ratio}, or \code{asp} parameters. Default
\code{FALSE}.}

\item{from_crs}{coordinate reference system of the data.}

\item{crs}{coordinate reference system to return}

\item{label_col}{Column name or id for a column with the text or labels to
pass to any text geom.}

\item{...}{Parameters passed to selected geom}
}
\value{
ggplot2 geom
}
\description{
Helper function to make a ggplot2 layer from data returned by
\code{\link[=get_location_data]{get_location_data()}}. For text geoms, the required aesthetic mapping is
set based on the name_col but values passed to mapping take precedence.
}
\details{
Supported geom function options:

Options for the geom parameter from the overedge package include:
\itemize{
\item "icon" (\link{layer_icon}),
\item "mapbox" (\link{layer_mapbox}),
\item "markers" (\link{layer_markers}),
\item "numbers" (\link{layer_numbers})
}

Options for the geom parameter from other packages include:
\itemize{
\item "textsf" (\link[geomtextpath:geom_textsf]{geomtextpath::geom_textsf})
\item "labelsf" (\link[geomtextpath:geom_textsf]{geomtextpath::geom_labelsf})
\item "text_repel" (\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel})
\item "label_repel" (\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel})
\item "mark" (\link[birdseyeview:layer_show_mark]{birdseyeview::layer_show_mark})
\item "location" (\link[birdseyeview:layer_show_location]{birdseyeview::layer_show_location})
\item "context" (\link[birdseyeview:layer_show_context]{birdseyeview::layer_show_context})
\item "pattern" (\link[ggpattern:geom-docs]{ggpattern::geom_sf_pattern})
}

Alternatively, use the "geom_fn" parameter to pass a function that returns a
ggplot2 layer to use instead of one of the preset geom functions.
}
\seealso{
\code{\link[ggplot2]{CoordSf}}

Other layer: 
\code{\link{layer_frame}()},
\code{\link{layer_icon}()},
\code{\link{layer_markers}()},
\code{\link{layer_mask}()},
\code{\link{layer_neatline}()},
\code{\link{layer_scaled}()}
}
\concept{layer}
