% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_markers.R
\name{layer_markers}
\alias{layer_markers}
\alias{layer_numbers}
\title{Add a marker layer to a map with or without numbered markers}
\usage{
layer_markers(
  data,
  mapping = NULL,
  geom = "sf",
  make = FALSE,
  groupname_col = NULL,
  group_meta = NULL,
  crs = NULL,
  number = FALSE,
  num_by_group = FALSE,
  num_style = NULL,
  suffix = NULL,
  sort = "dist_xmin_ymax",
  desc = FALSE,
  fn = NULL,
  ...
)

layer_numbers(
  data,
  mapping = NULL,
  geom = "label",
  make = FALSE,
  groupname_col = NULL,
  style = "roundrect",
  size = 5,
  sort = "dist_xmin_ymax",
  num_by_group = FALSE,
  num_style = NULL,
  suffix = NULL,
  desc = FALSE,
  fn = NULL,
  crs = NULL,
  label.size = 0,
  label.padding = ggplot2::unit(size/10, "lines"),
  label.r = label.padding * 1.5,
  hjust = 0.5,
  vjust = 0.5,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of
\code{\link[=get_location_data]{get_location_data()}}}

\item{make}{If \code{TRUE}, pass data to \link{make_markers}.}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers; Default: \code{NULL}}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_markers]{layer_markers()}} (not
currently supported)}

\item{num_by_group}{If TRUE, numbers are added by group based on
groupname_col.}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph",
"roman", "Roman"}

\item{suffix}{Character to appended to "number" column. (e.g. "." for "1." or
":" for "1:"). Can also be a character vector with the same length as the
number column.}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}

\item{fn}{Function to apply to data before results; gives warning if data is
grouped; Default: \code{NULL}}

\item{...}{Additional parameters passed to \code{\link[=layer_location_data]{layer_location_data()}}}

\item{style}{Style of number markers to map; defaults to "roundrect".}

\item{size}{Marker size, Default: 5}
}
\value{
ggplot2 layers
}
\description{
If get is \code{TRUE}, groupname_col, group_meta, crs, and fn is all passed on to
make_markers.
}
\details{
The number parameter is not currently supported so the number_col parameter
is not implemented.
}
\examples{
\dontrun{
if (interactive()) {
  ggplot2::ggplot() +
    layer_markers(
      data = mapbaltimore::parks,
      groupname_col = "park_district"
    )
}
}
}
