% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mask.R
\name{layer_mask}
\alias{layer_mask}
\title{Create a mask layer based on a simple feature object}
\usage{
layer_mask(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = NULL,
  fill = "white",
  color = NA,
  alpha = 0.5,
  mask = NULL,
  neatline = FALSE,
  ...
)
}
\arguments{
\item{data}{sf, sfc, or bbox object. If dist, diag_ratio, and/or asp are provided, data is adjusted to set the boundaries of the mask. If data is not provided, \code{mask} is required.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Units are passed to
\code{units::set_units()} and then converted to units matching GDAL; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, get_asp returns the same value without modification.}

\item{crs}{Coordinate reference system of bounding box to return}

\item{fill}{mask fill color; defaults to "white"}

\item{color}{mask edge color; defaults to NA}

\item{alpha}{mask alpha/transparency; defaults to 0.5}

\item{mask}{\code{bbox} object to define the edge of the mask.
\code{diag_ratio}, \code{dist}, and \code{asp} parameters are ignored if a
\code{mask} is provided. defaults to NULL}

\item{neatline}{If TRUE, add \code{layer_neatline} with \code{expand = TRUE}; defaults to FALSE.}

\item{...}{Additional parameters to pass to \code{\link[ggplot2]{geom_sf}}}
}
\value{
\code{\link[ggplot2]{geom_sf}} function.
}
\description{
Returns a mask for an area or areas as an  \code{sf} object.
}
