% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_neatline.R
\name{layer_neatline}
\alias{layer_neatline}
\alias{set_map_limits}
\title{Set map limits to sf or bbox object with optional adjustments and panel border}
\usage{
layer_neatline(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = NULL,
  color = "black",
  size = 1,
  linetype = "solid",
  expand = FALSE,
  hide_grid = TRUE,
  label_axes = "----",
  ...
)
}
\arguments{
\item{data}{sf or bbox class object}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{Coordinate reference system to use for
\code{\link[sf]{coord_sf}}.}

\item{color}{color of panel border, Default: 'black'}

\item{size}{size of panel border, Default: 1}

\item{linetype}{line type of panel border, Default: 'solid'}

\item{expand}{Default FALSE. If TRUE, use scale_y_continuous and
scale_x_continuous to expand map extent to provided parameters.}

\item{hide_grid}{logical. If TRUE, hide major grid lines. Default: TRUE}

\item{label_axes}{label_axes passed to \code{\link[sf]{coord_sf}}, Default:
'----' which hides axes labels.}

\item{...}{Additional parameters to pass to \code{\link[sf]{coord_sf}}.}
}
\value{
\code{ggplot2::coord_sf()} function with xlim and ylim parameters
}
\description{
Set limits for a map to the bounding box of an x using
\code{\link[sf]{coord_sf}}. Optionally, adjust the x size by applying a
buffer and/or adjust the aspect ratio of the limiting bounding box to match a
set aspect ratio.
}
\examples{
\dontrun{
if(interactive()){
 library(ggplot2)

  nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
  ggplot() +
    geom_sf(data = nc) +
    layer_neatline(data = nc[1,], asp = 1, color = "red")

  }
}
}
\seealso{
\code{\link[ggplot2]{CoordSf}},\code{\link[ggplot2]{scale_continuous}}
}
