% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_scaled.R
\name{layer_scaled}
\alias{layer_scaled}
\title{Create a ggplot2 layer scaled to a paper and orientation for a location}
\usage{
layer_scaled(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = NULL,
  scale = NULL,
  series = NULL,
  standard = NULL,
  paper = NULL,
  orientation = NULL,
  clip = FALSE
)
}
\arguments{
\item{data}{\code{sf} or \code{bbox} class object}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Units are passed to
\code{units::set_units()} and then converted to units matching GDAL; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system of bounding box to return}

\item{scale}{Scale name}

\item{series}{Map series}

\item{standard}{USGS, Engineering, or Architectural}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}
}
\description{
Uses \link{layer_neatline}, \link{standard_scales}, and \link{get_standard_scale}.
}
\seealso{
Other layer: 
\code{\link{layer_location_data}()},
\code{\link{layer_mask}()},
\code{\link{layer_neatline}()},
\code{\link{layer_show_location}()}
}
\concept{layer}
