% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{make_filename}
\alias{make_filename}
\title{Make file name and path with optional label, prefix, or postfix}
\usage{
make_filename(
  name = NULL,
  label = NULL,
  filetype = NULL,
  filename = NULL,
  path = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{name}{Name to make file name converted to snake case with
\code{\link[janitor]{make_clean_case}}, e.g. "Residential zoning map"
becomes "residential_zoning_map"}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor]{make_clean_case}}. The label is designed to identify
the area or other shared characteristics across multiple data files, maps,
or plots.}

\item{filetype}{File type or extension.}

\item{filename}{File name; if file name is \code{NULL}, name and file type are both
required.}

\item{path}{Path to file or data directory.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}
}
\description{
A helper function to create consistent file names for plots created with
\code{\link[=ggsave_ext]{ggsave_ext()}} or data files exported with \code{\link[=write_sf_ext]{write_sf_ext()}}.
}
