% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_grid.R
\name{make_location_grid}
\alias{make_location_grid}
\title{Make a grid over a location bounding box}
\usage{
make_location_grid(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = "meter",
  crs = NULL,
  cols = NULL,
  rows = NULL,
  gutter = 0,
  desc = FALSE,
  n = NULL,
  cellsize = NULL,
  what = NULL,
  style = "rect",
  .id = "id"
)
}
\arguments{
\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}. Required.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{crs}{Coordinate reference system of bounding box to return}

\item{cols, rows}{Used to set n if either are not \code{NULL}; defaults to \code{NULL}.
row and id are added as columns to the grid if they are provided.}

\item{gutter}{Distance in units between each column cell; gutter effectively
serves as a margin as the negative buffer is applied to all cells
(including those at the edges of the grid).}

\item{desc}{If TRUE, reverse standard order of cell id numbering; defaults
\code{FALSE}}

\item{n}{If n is NULL and square is \code{TRUE}, the grid is set automatically to
be 10 cells wide, Default: \code{NULL}}

\item{cellsize}{target cellsize}

\item{what}{"polygons", "corners", "centers"; set to centers automatically if
style is "circle", "circle_offset" but a buffer is applied to return
circular polygons.}

\item{style}{Style of cell to return with options including "rect", "square",
"hex", "flat_top_hex", "circle", "circle_offset"}

\item{.id}{A name to use for the cell id column.}
}
\description{
Create a grid with an id column and optionally a set number of columns and
rows. This documentation is incomplete the function may change.
}
\examples{
# Make a 4 by 5 grid covering a mile buffer around a location
make_location_grid(
  location = mapbaltimore::council_districts[1, ],
  dist = 1,
  unit = "mile",
  cols = 4,
  rows = 5
)

# Make a 2 by 2 grid across a location with a 1000 meter gutter between each cell
make_location_grid(
  location = mapbaltimore::baltimore_city,
  dist = 500,
  unit = "meter",
  cols = 2,
  rows = 2,
  gutter = 1000
)
}
\seealso{
\link[sf:st_make_grid]{sf::st_make_grid}
}
