% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{make_photo_popup}
\alias{make_photo_popup}
\title{Make a photo popup}
\usage{
make_photo_popup(photos, popup = TRUE)
}
\arguments{
\item{popup}{If \code{TRUE}, add a popup image to a leaflet map; defaults \code{TRUE}.}
}
\description{
Utility function for \link{mapview_flickr} and \link{mapview_exif}
}
