% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{mapview_ext}
\alias{mapview_ext}
\alias{mapview_col}
\title{Use mapview to interactively explore data}
\usage{
mapview_col(x, nm = "data", col = NULL, ...)
}
\arguments{
\item{x}{a \code{Raster*} or \code{Spatial*} or \code{Satellite} or
\code{sf} or \code{stars} object or a list of any combination of those. Furthermore,
  this can also be a \code{data.frame}, a \code{numeric vector} or a
\code{character string} pointing to a tile image folder or file on disk.
  If missing, a blank map will be drawn. A value of NULL will return NULL.}

\item{nm}{Character vector. If x is a sf list, filter the mapview data to
those named in nm. If NULL, mapview displays the first item in the sf list;
defaults to "data".}

\item{col}{Column name passed to zcol parameter, Default: NULL}

\item{...}{
  Arguments passed on to \code{\link[mapview:mapView]{mapview::mapview}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
A wrapper for \link[mapview:mapView]{mapview::mapview}
}
\seealso{
\code{\link[mapview]{mapView}}
}
