% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paper_sizes}
\alias{paper_sizes}
\title{Common paper sizes}
\format{
A data frame with 85 rows and 9 variables:
\describe{
\item{\code{name}}{Name of paper}
\item{\code{series}}{Series}
\item{\code{standard}}{Standard}
\item{\code{size}}{Size in series}
\item{\code{units}}{Units ("in" or "mm") for dimensions}
\item{\code{width}}{Width in units}
\item{\code{height}}{Height in units}
\item{\code{asp_portrait}}{Aspect ratio (portrait)}
\item{\code{asp_landscape}}{Aspect ratio (landscape)}
\item{\code{asp_text}}{Aspect ratio ("width:height")}
}
}
\usage{
paper_sizes
}
\description{
Reference table of common paper sizes used for \code{get_paper()} and
\code{ggsave_ext()} functions. Derived from
\href{https://github.com/visioguy/PaperSizes/}{visioguy/PaperSizes} repo.
}
\keyword{datasets}
