% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_group.R
\name{scale_group_data}
\alias{scale_group_data}
\alias{group_data_pal}
\alias{get_group_data_pal_scale}
\title{Create discrete fill and color scales for grouped data}
\usage{
scale_group_data(
  ...,
  data,
  col = NULL,
  palette = NULL,
  n = NULL,
  direction = 1,
  na.value = "grey50",
  drop = FALSE,
  limits = NULL,
  aesthetics = "fill"
)

group_data_pal(data, palette = NULL, col = NULL, n = NULL, direction = 1)

get_group_data_pal_scale(data, col = NULL, palette = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual}}
  \describe{
    \item{\code{values}}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
  }}

\item{data}{Data to use when generating scale or palette.}

\item{col}{Grouping column found in data to use in generating scale or
palette; defaults to \code{NULL}.}

\item{palette}{Name of palette as a string. Must be on the form
packagename::palettename.}

\item{n}{Number of colors desired. If omitted, returns complete palette.}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1} the palette will be reversed.}

\item{na.value}{The aesthetic value to use for missing (\code{NA}) values}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}
\description{
Designed for use with layer_group_data. group_data_pal generates palettes
that are passed to \link[ggplot2:scale_manual]{ggplot2::scale_fill_manual} and
\link[ggplot2:scale_manual]{ggplot2::scale_color_manual}.
}
\seealso{
\code{\link[scales]{viridis_pal}}
\code{\link[paletteer]{paletteer_d}}
}
