% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_join_by_geom_type}
\alias{set_join_by_geom_type}
\title{Set join function based on geometry type}
\usage{
set_join_by_geom_type(x, join = NULL)
}
\arguments{
\item{join}{geometry predicate function; defaults to \code{NULL}, set to
\link[sf:geos_binary_pred]{sf::st_intersects} if key_list contains only POLYGON or MULTIPOLYGON objects
or \link[sf:st_nearest_feature]{sf::st_nearest_feature} if key_list contains other types.}
}
\description{
Set join function based on geometry type
}
