% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{set_overedge_options}
\alias{set_overedge_options}
\title{Set overedge packge options}
\usage{
set_overedge_options(
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data_package = NULL,
  data_filetype = NULL,
  from_crs = NULL,
  crs = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{dist, diag_ratio, unit, asp, data_package, data_filetype, from_crs, crs}{options to set, e.g. "crs = 2804" with
\code{pkg = "overedge"} to set "overedge.crs" to 2804.}

\item{overwrite}{If \code{TRUE}, overwrite any existing option value.}
}
\description{
Can set options for package, diag_ratio, dist, asp, or crs
}
