% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox.R
\name{sf_bbox_to_lonlat_query}
\alias{sf_bbox_to_lonlat_query}
\title{Convert a bounding box to lon/lat query string}
\usage{
sf_bbox_to_lonlat_query(
  bbox,
  coords = c("longitude", "latitude"),
  crs = 4326,
  lonlat = TRUE
)
}
\arguments{
\item{bbox}{bounding box}

\item{coords}{query column with lon/lat coordinates}

\item{crs}{coordinate reference system to use for query; default 4326}

\item{lonlat}{Default TRUE. If FALSE, coords is assumed to be lat/lon order instead of lon/lat.}
}
\description{
Take a bounding box and return a SQL style query string with lon/lat values.
}
