% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox.R
\name{sf_bbox_to_sf}
\alias{sf_bbox_to_sf}
\title{Convert a bounding box into simple feature object}
\usage{
sf_bbox_to_sf(bbox)
}
\arguments{
\item{bbox}{bbox object}
}
\value{
sf object
}
\description{
Take a bbox and return an sf object.
}
\seealso{
\itemize{
\item \code{\link[sf]{st_as_sf}},\code{\link[sf]{st_as_sfc}}
\item \code{\link[sfx]{st_extent}}
}
}
