% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox.R
\name{sf_bbox_to_wkt}
\alias{sf_bbox_to_wkt}
\title{Convert a bounding box to well known text (wkt)}
\usage{
sf_bbox_to_wkt(bbox)
}
\arguments{
\item{bbox}{bbox object}
}
\value{
well known text
}
\description{
Take a bbox and return well known text for use in the wkt_filter parameter of
\code{\link[sf]{read_sf}}.
}
\seealso{
\code{\link[sf]{st_as_text}},\code{\link[sf]{st_as_sfc}}
}
