% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_df.R
\name{sf_to_df}
\alias{sf_to_df}
\title{Convert simple feature object to data frame}
\usage{
sf_to_df(
  x,
  crs = 4326,
  coords = c("lon", "lat"),
  geometry = "centroid",
  keep_all = TRUE
)
}
\arguments{
\item{x}{sf or sfc object}

\item{crs}{coordinate reference system, Default: 4326}

\item{coords}{coordinate columns used if geometry is 'centroid' or 'point';
Default: c("lon", "lat")}

\item{geometry}{type of geometry to include in data frame. options include
"drop", "wkt", "centroid", "point", Default: 'centroid'}

\item{keep_all}{If FALSE, drop all columns other than those named in coords,
Default: TRUE}
}
\value{
a data frame with geometry dropped or converted to wkt or coordinates
for the centroid or point on surface
}
\description{
Helper function to convert a simple feature object to data frame by dropping
geometry, converting geometry to well known text, or (if the geometry type is
not POINT) getting coordinates for a centroid or point on surface. If an sfc
object is provided, sf_to_sfc provides coordinates but the "drop" geometry
option is not supported.
}
\seealso{
\code{\link[sf]{st_coordinates}}
}
