% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_adj.R
\name{st_bbox_asp}
\alias{st_bbox_asp}
\title{Get bounding box adjusted to match aspect ratio}
\usage{
st_bbox_asp(x = NULL, asp = NULL)
}
\arguments{
\item{x}{\code{sf} or bbox object}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}
}
\value{
\code{bbox} object
}
\description{
Get bbox from sf or bbox object adjusted to match an aspect ratio
}
\details{
Takes an area as an  \code{sf} or \code{bbox} object and returns a bounding
box that matches the aspect ratio provided to \code{asp} and contains the
area or bounding box provided. Common aspect ratios include "1:1" (1), "4:6"
(0.666), "8.5:11", "16:9" (1.777). The asp parameter supports both numeric
values and character strings with ratios matching the format of
"width:height".
}
