% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bearing.R
\name{st_bearing}
\alias{st_bearing}
\title{Get bearing of a simple feature object with LINESTRING geometry}
\usage{
st_bearing(x, dir = FALSE, crs = 4326)
}
\arguments{
\item{x}{A \code{sf} object.}

\item{dir}{Logical indicator whether to include direction in bearing; If
\code{FALSE}, return the absolute (positive) bearing value. If \code{TRUE}, return
negative and positive bearing values. Default: \code{FALSE}.}

\item{crs}{Coordinate reference system passed to
\code{\link[sf]{st_coordinates()}} (must be geographic not projected).
Default: 4326.}
}
\description{
Get bearing of a simple feature object with LINESTRING geometry
}
\examples{
\dontrun{
if (interactive()) {
  library(ggplot2)

  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
  nc_pts <- sf::st_centroid(nc)
  nc_line <-
    sf::st_as_sf(
      sf::st_cast(
        sf::st_union(c(nc_pts[1, ]$geometry, nc_pts[30, ]$geometry)),
        "LINESTRING"
      )
    )

  ggplot() +
    geom_sf(data = nc) +
    geom_sf(data = nc_pts) +
    geom_sf(data = st_bearing(nc_line), aes(color = bearing))
}
}
}
\seealso{
\code{\link[lwgeom]{st_startpoint}},\code{\link[lwgeom]{st_endpoint}}
\code{\link[geosphere]{bearing}}
}
