% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cast_ext.R
\name{st_cast_ext}
\alias{st_cast_ext}
\title{Cast geometry to another type}
\usage{
st_cast_ext(x, to = "POINT", ...)
}
\arguments{
\item{x}{A \code{sf} or \code{sfc} object to cast to another type.}

\item{to}{character; target type, if missing, simplification is tried; when \code{x} is of type \code{sfg} (i.e., a single geometry) then \code{to} needs to be specified.}

\item{...}{ignored}
}
\description{
Wrapper for \code{\link[sf:st_cast]{sf::st_cast()}} that currently only supports casting MULTIPOLYGON
to POLYGON or MULTIPOLYGON or POLYGON to POINT.
}
\seealso{
\code{\link[sf]{st_cast}}
}
