% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_clip.R
\name{st_clip}
\alias{st_clip}
\title{Clip a simple feature or bounding box to a side or corner}
\usage{
st_clip(
  x,
  clip = NULL,
  keep = NULL,
  flip = FALSE,
  dist = NULL,
  diag_ratio = NULL
)
}
\arguments{
\item{x}{\code{sf} or \code{bbox} object to clip}

\item{clip}{Character string describing the part of the area to clip or
remove. Options include c("top", "right", "bottom", "left", "topright",
"bottomright", "bottomleft", "topleft"). If NULL, the area is not clipped
and a full edge can be returned.}

\item{keep}{Alternate way of defining clip (by naming the section to keep).}

\item{flip}{Logical. Default FALSE. If TRUE, than the clip area
is kept instead of removed. If keep is provided, flip is automatically set to TRUE.}

\item{dist}{Numeric. Distance to use for the edge. Default NULL
meters. Use negative values for an inside edge or positive numbers for an
outside edge.}

\item{diag_ratio}{Alternate way to define edge distance.}
}
\value{
\code{sf} object clipped based on parameters
}
\description{
Clip based on the corner of the object bounding box.
}
