% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dist.R
\name{st_dist}
\alias{st_dist}
\title{Get distance from a simple feature object to another object or bounding box
corner}
\usage{
st_dist(
  from,
  to,
  by_element = TRUE,
  units = NULL,
  drop = FALSE,
  keep_all = TRUE
)
}
\arguments{
\item{from}{A sf, sfc, or bbox object.}

\item{to}{A sf, sfc, or bbox object or length 2 character vector. If sf or
sfc, must have 1 feature or the same number of features as from (if
by_element is TRUE). If to is a character vector it must represent a valid
xy pair using the following options: "xmin", "ymin", "xmax", "ymax",
"xmid", "ymid".}

\item{by_element}{logical; if \code{TRUE}, return a vector with distance between the first elements of \code{x} and \code{y}, the second, etc. if \code{FALSE}, return the dense matrix with all pairwise distances.}

\item{units}{Unit to convert distance to, Default: NULL}

\item{keep_all}{If FALSE, return dist but not original sf, Default: TRUE}
}
\description{
Get distance from a simple feature object to another object or bounding box
corner
}
