% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_inscribed_square.R
\name{st_inscribed_square}
\alias{st_inscribed_square}
\title{Create an inscribed square in a simple feature object}
\usage{
st_inscribed_square(x, rotate = 0)
}
\arguments{
\item{x}{object of class sf}

\item{rotate}{numeric; degrees to rotate square geometry, Default: 0}
}
\description{
This function wraps \code{sf::st_inscribed_circle()} but limits the
circle to 1 segment per quadrant (\code{nQuadSegs} = 1) and then rotates the resulting geometry 45
degrees to provide a (mostly) inscribed square. A different rotation value
can be provided to change the orientation of the shap, e.g. rotate = -45 to
return a diamond shape.
}
\examples{
\dontrun{
if(interactive()){
 library(ggplot2)

 nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
 nc_squares <- st_inscribed_square(sf::st_transform(nc, 2264))

 ggplot() +
  geom_sf(data = nc) +
  geom_sf(data = nc_squares, fill = NA, color = "red")
 }
}
}
\seealso{
\code{\link[sf]{geos_unary}}
}
