% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_scale_rotate.R
\name{st_scale_rotate}
\alias{st_scale_rotate}
\title{Scale or rotate a simple feature object}
\usage{
st_scale_rotate(x, scale = 1, rotate = 0)
}
\arguments{
\item{x}{object of class sf}

\item{scale}{numeric; scale factor, Default: 1}

\item{rotate}{numeric; degrees to rotate (-360 to 360), Default: 0}
}
\description{
Scale or rotate a simple feature object using affine transformations
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(ggplot2)
 library(overedge)

 nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
 nc_rotated <- st_scale_rotate(nc, scale = 0.5, rotate = 15)

 ggplot() +
  geom_sf(data = nc) +
  geom_sf(data = nc_rotated, fill = NA, color = "red")
 }
}
}
\seealso{
\code{\link[sf]{geos_unary}}
}
