% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{str_pad_digit}
\alias{str_pad_digit}
\title{Pad a string with digits}
\usage{
str_pad_digit(string, pad = "0", side = "left", width = NULL)
}
\arguments{
\item{string}{A character vector.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{side}{Side on which padding character is added (left, right or both).}

\item{width}{Minimum width of padded strings.}
}
\description{
Pad a string with digits
}
