% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{str_prefix}
\alias{str_prefix}
\title{Add prefix and postfix to string}
\usage{
str_prefix(
  prefix = NULL,
  string = NULL,
  postfix = NULL,
  sep = "_",
  clean = TRUE,
  pad = "0",
  width = NULL
)
}
\arguments{
\item{prefix}{Character string or character vector to add to string parameter
as a prefix.}

\item{string}{A single string that the attach prefix or postfix is added to.}

\item{postfix}{Character string or character vector to add to string
parameter as a postfix.}

\item{sep}{Separator character passed as the collapse parameter of \code{\link[=paste]{paste()}}.}

\item{clean}{If \code{TRUE}, prefix, postfix, and string are all converted to
snake case with \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{width}{Minimum width of padded strings.}
}
\description{
Prefix and postfix can include more than one value that are added in the same
order provided. String must be a single character string.
}
