% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{str_remove_filetype}
\alias{str_remove_filetype}
\title{Remove file type from string}
\usage{
str_remove_filetype(x = NULL, filetype = NULL)
}
\arguments{
\item{x}{String}

\item{filetype}{File type string}
}
\description{
Remove file type from string
}
