% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ext.R
\name{theme_ext}
\alias{theme_ext}
\alias{theme_text}
\alias{theme_margin}
\alias{theme_legend}
\title{Modify the text, margins, or legend for a ggplot theme}
\usage{
theme_text(
  font_family = NULL,
  color = "black",
  geom_text = TRUE,
  method = NULL
)

theme_margin(
  margin = "standard",
  paper = NULL,
  orientation = NULL,
  dist = NULL,
  unit = "in",
  plot_width = NULL,
  header = 0,
  footer = 0,
  fill = NA,
  color = NA,
  size = 0,
  method = NULL
)

theme_legend(
  position,
  margin = 10,
  unit = "pt",
  inset = TRUE,
  bgcolor = "white",
  justification = NULL,
  method = NULL
)
}
\arguments{
\item{font_family}{Font family, Default: 'Helvetica' If NULL, font_family is
pulled from current set theme which is helpful for resetting all text
families to the theme.}

\item{color}{color for text elements (passed to \code{\link[ggplot2]{element_text}} by
theme_text), plot.background (passed to \code{\link[ggplot2]{element_rect}} by
theme_margin). Default: NA}

\item{geom_text}{If TRUE, update text family for \code{\link[ggplot2]{geom_text}}, \code{\link[ggplot2]{geom_sf_text}},
\code{\link[ggplot2]{geom_label}}, and \code{\link[ggplot2]{geom_sf_label}} to match \code{font_family} and color. If FALSE,
make no changes to the theme. Default: TRUE}

\item{method}{ggplot2 function to use in changing text "set", "update", or
"replace", Default: 'set'}

\item{margin}{margin distance, a margin style supported by get_margins or a
margin object, Default: 10}

\item{paper}{Paper, Default: 'letter'}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'}

\item{dist}{margin distance (single value used to all sides), Default: NULL}

\item{unit}{margin units, Default: 'pt'}

\item{plot_width}{plot or map width in units. If \code{paper} and \code{plot_width} are
provided, margins are half the distance between the two evenly distributed.
This is not tested and may not work with all page sizes/orientations.}

\item{header}{header height; defaults to 0}

\item{footer}{footer height; defaults to 0}

\item{fill}{fill for plot.background passed to \code{\link[ggplot2]{element_rect}}
Default: NA}

\item{size}{size number (only used for "ISO" and "JIS" series). Standard,
series, and size may all be required to return a single paper using these
parameters.}

\item{position}{position type or a a two-element numeric vector}

\item{inset}{If TRUE and position is "topleft", "bottomleft", "topright", or
"bottomright", place the legend in an inset position. Default: TRUE}

\item{bgcolor}{fill color for legend background, Default: 'white'}

\item{justification}{If NULL, justification is set to "center", Default: NULL}
}
\description{
Modify the text, margins, or legend for a ggplot theme
}
\examples{
\dontrun{
if (interactive()) {
  theme_margin(paper = "letter")

  theme_margin(margin = "narrow", paper = "letter")
}
}
}
\seealso{
\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{theme_get}},\code{\link[ggplot2]{update_geom_defaults}
}
