% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_text.R
\name{theme_margin}
\alias{theme_margin}
\title{Set, update, or replace margins for ggplot theme}
\usage{
theme_margin(
  margin = "standard",
  paper = NULL,
  orientation = NULL,
  dist = NULL,
  unit = "in",
  plot_width = NULL,
  fill = NA,
  color = NA,
  size = 0,
  method = NULL
)
}
\arguments{
\item{paper}{Paper, Default: 'letter'}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default: 'portrait'}

\item{fill}{fill for plot.background passed to ggplot2::element_rect(), Default: NA}

\item{color}{color for plot.background passed to ggplot2::element_rect(), Default: NA}

\item{method}{ggplot2 function to use in changing text "set", "update", or
"replace", Default: 'set'}
}
\description{
Set, update, or replace margins for ggplot theme
}
\examples{
\dontrun{
if (interactive()) {
  ggplot() +
    layer_loc
  theme_margin(paper = "letter")

  theme_margin(margin = "narrow", paper = "letter")
}
}
}
\seealso{
\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
}
