% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_text.R
\name{theme_text}
\alias{theme_text}
\title{Set, update, or replace ggplot theme and geom text family}
\usage{
theme_text(font_family = "Helvetica", method = "set", label = FALSE)
}
\arguments{
\item{font_family}{Font family, Default: 'Helvetica' If NULL, font_family is
pulled from current set theme which is helpful for resetting all text
families to the theme.}

\item{method}{ggplot2 function to use in changing text "set", "update", or
"replace", Default: 'set'}

\item{label}{If TRUE, update default text family for geom_text, geom_sf_text,
geom_label, and geom_sf_label. If FALSE, update default text family for
geoms to match current theme. Default: TRUE}
}
\description{
Set, update, or replace ggplot theme and geom text family
}
\seealso{
\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{theme_get}},\code{\link[ggplot2]{update_geom_defaults}}
}
