% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_counties}
\alias{us_counties}
\title{U.S. County boundaries (1:5 mi scale, bbox and wkt)}
\format{
A data frame with 3220 rows and 7 variables:
\describe{
\item{\code{name}}{County name}
\item{\code{geoid}}{County GeoID}
\item{\code{countyfp}}{County FIPS}
\item{\code{statefp}}{State FIPS}
\item{\code{abb_state}}{State abbreviation (USPS)}
\item{\code{bbox}}{Bounding box}
\item{\code{wkt}}{Well known text}
}
}
\usage{
us_counties
}
\description{
U.S. County boundaries data downloaded with \link[tigris:counties]{tigris::counties}.
}
\keyword{datasets}
