% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{U.S. State boundaries (1:20 mi scale, bbox and wkt)}
\format{
A data frame with 52 rows and 6 variables:
\describe{
\item{\code{name}}{State name}
\item{\code{geoid}}{State GeoID}
\item{\code{statefp}}{State FIPS}
\item{\code{abb}}{State abbreviation (USPS)}
\item{\code{bbox}}{Bounding box}
\item{\code{wkt}}{Well known text}
}
}
\usage{
us_states
}
\description{
U.S. State boundaries data downloaded with \link[tigris:states]{tigris::states}.
}
\keyword{datasets}
