% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_point}
\alias{as_point}
\alias{as_points}
\title{Convert a simple feature object or a numeric string to a single POINT (sfg)
or POINT geometry set (sfc)}
\usage{
as_points(...)

as_point(...)
}
\description{
Works with sf, sfc, and bbox objects using \link[sf:geos_unary]{sf::st_centroid}. Works with
\link{sf_bbox_point}
}
