% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_bbox}
\alias{as_sfc}
\alias{as_sf_list}
\title{Convert an object to a simple feature or bounding box object}
\usage{
as_sf(x, crs = NULL, sf_col = "geometry", ...)

as_bbox(x, crs = NULL, ...)

as_sfc(x, crs = NULL, ...)

as_sf_list(x, nm = "data", col = NULL, crs = NULL)
}
\arguments{
\item{x}{A \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, \code{sp}, or data frame object that
can be converted into a simple feature or bounding box object. \code{\link[=as_bbox]{as_bbox()}}
can also convert a vector with xmin, ymin, xmax, and ymax values.
\link{as_sf_list} only supports sf objects or a data frames with a sf list
column named "data" (typically created by using \code{\link[dplyr:group_nest]{dplyr::group_nest()}} on an
sf object.}

\item{crs}{Coordinate reference system for \code{sf}, \code{bbox}, \code{sfc} or \code{sf} list object to
return.}

\item{...}{Additional parameters passed to \code{\link[sf:st_bbox]{sf::st_bbox()}} when calling
\code{\link[=as_bbox]{as_bbox()}} or passed to \code{\link[sf:sf]{sf::st_sf()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}, or \code{\link[=df_to_sf]{df_to_sf()}}
for \code{\link[=as_sf]{as_sf()}} (depending on class of x)}

\item{nm}{For \link{as_sf_list}, name(s) for sf list; defaults to "data". If col is provided, the
values of the grouping column are used as names.}

\item{col}{For \link{as_sf_list}, the name of the column used to group data if x is a sf object or used
to group and nest data before passing to x.}
}
\description{
Both functions will pass a NULL value without returning an error. If a POINT
or MULTIPOINT object is passed to \link{as_bbox} a 0.00000001 meter buffer is
applied.
}
