% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sf.R
\name{check_sf}
\alias{check_sf}
\alias{check_bbox}
\alias{check_sf_list}
\alias{check_sf_same_crs}
\title{Check the class and attributes of a simple feature object}
\usage{
check_sf(x, ext = FALSE)

check_bbox(x)

check_sf_list(x)

check_sf_same_crs(x, y)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by check_sf)}

\item{y}{An sf object or a character or numeric object supported by
\code{\link[sf]{st_crs}} that can be compared to x. (used by
check_sf_same_crs)}
}
\description{
Check the class and attributes of a simple feature object
}
\details{
\itemize{
\item check_sf: check if x it a \code{sf} class object.
\item check_bbox: check if x is a \code{bbox} class object.
\item check_sf_list: check if x is a named list of \code{sf} class objects.
}
}
