% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_buffer_ext.R
\name{convert_dist_units}
\alias{convert_dist_units}
\title{Convert distances between different units}
\usage{
convert_dist_units(dist, from = NULL, to = "meter")
}
\arguments{
\item{dist}{Numeric or units}

\item{from}{Existing unit for dist, Default: NULL. If dist is a units object, the numerator is used as from_unit}

\item{to}{Unit to convert distance to, Default: 'meter'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Convert distances between different units
}
