% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_icon.R
\name{geom_sf_icon}
\alias{geom_sf_icon}
\title{Draw SVG icons for simple feature objects}
\usage{
geom_sf_icon(
  data = NULL,
  iconname_col = "icon",
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  ...
)
}
\arguments{
\item{data}{sf object. If the geometry type for data is POINT, the object is
used as is. If not, the object is converted to POINT using
\code{\link[sf]{st_centroid()}}.}

\item{iconname_col}{The column name in the input data to use as the icon
name. If the name matches multiple icons, the first match from \code{map_icons}
is used. You may provide a px or source value to select a different match
if needed but, in that case, all icons must use the same px or source
value. Note that the icon column should not be mapped with \code{aes()}.}

\item{icon}{Icon name. Default NULL. If \code{icon} is provided, \code{iconname_col} is
not used.}

\item{px}{Icon size in pixels. See \code{map_icons$px} for supported options.
Optional but may be necessary to differentiate icons with duplicate names.}

\item{source}{Icon source. See \code{map_icons$repo} for supported options.
Optional but may be required to differentiate icons with duplicate names.}

\item{svg}{Optional. Custom file path or URL with SVG to pass to \code{svg}
parameter for \code{ggsvg::geom_point_svg()}.  If \code{icon} is provided, \code{svg} is
not used.}

\item{...}{Additional parameters to ggsvg::geom_point_svg()}
}
\description{
Use the \code{\link[ggsvg]{geom_point_svg()}} function to plot
icons using the centroids from the input simple feature object to set the
icon location.
}
\examples{
\dontrun{
if (interactive()) {
  library(sf)
  library(ggplot2)
  library(overedge)

  nc <- st_read(system.file("shape/nc.shp", package = "sf"))
  ggplot() +
    geom_sf_icon(data = nc, icon = "point-start", size = 10)


  nc$icon <- rep(c("1", "2", "3", "4"), nrow(nc) / 4)
  ggplot() +
    geom_sf_icon(data = nc, size = 5)
}
}
}
\seealso{
\code{\link[ggsvg]{geom_point_svg}}
\code{\link[overedge]{map_icons}}
}
