% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_adj.R
\name{get_asp}
\alias{get_asp}
\title{Get aspect ratio from string or based on specific paper and margins}
\usage{
get_asp(
  asp = NULL,
  paper = NULL,
  orientation = NULL,
  margin = NULL,
  unit = NULL,
  block_asp = FALSE
)
}
\arguments{
\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, get_asp returns the same value without modification.}

\item{paper}{Paper, Default: 'letter'}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'}

\item{margin}{margin style (options include "extrawide", "wide", "standard",
"narrow", "none"), Additional "auto" option to generate margin based on
line length is planned but not yet implemented. Default: NULL (equivalent to "none")}

\item{unit}{unit for margin distance, Default: 'in'}

\item{block_asp}{If TRUE, get aspect ratio of the map/plot area (not including the margins); defaults to FALSE.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get aspect ratio from string or based on specific paper and margins
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  get_asp("1:2")

  get_asp(11 / 17)

  get_asp(paper = "letter")
}
}
}
