% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dir.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Check data directory and create if needed}
\usage{
get_data_dir(path = NULL, package = "overedge")
}
\arguments{
\item{path}{Path to directory to use as data directory.}

\item{package}{Package name; defaults to "overedge"}
}
\description{
A utility function that wraps \link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir}.
}
