% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_data.R
\name{get_esri_data}
\alias{get_esri_data}
\title{Get data from an ArcGIS FeatureServer or MapServer for a location}
\usage{
get_esri_data(
  location = NULL,
  url,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = NULL,
  where = NULL,
  coords_col = NULL,
  lonlat = TRUE,
  locationname_col = NULL,
  locationname,
  ...
)
}
\arguments{
\item{location}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \code{\link[esri2sf]{esri2sf}} or
\code{\link[esri2sf]{esri2df}} functions.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "m", "ft", "km", and
"mi". Equivalent values (e.g. "mile" instead of "mi") are also supported.
Units are passed to \code{units::set_units()} and then converted to units
matching GDAL; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{coordinate reference system to use for query; default 4326}

\item{where}{where query string passed to esri2sf, Default: NULL}

\item{coords_col}{coordinate columns, e.g. c("longitude", "latitude")}

\item{lonlat}{If FALSE, assume coords_col is in lat/lon order; defaults to
TRUE,}

\item{locationname_col}{name of ArcGIS FeatureServer or MapServer column with
location names for features}

\item{locationname}{location name}
}
\description{
Wraps the \code{\link[esri2sf]{esri2sf}} and \code{\link[esri2sf]{esri2df}}
function to download an ArcGIS FeatureServer or MapServer. Supports spatial
filtering with bounding box based on location and filtering by location name
(if location name column is provided).
}
\seealso{
\code{\link[esri2sf]{esri2sf}}
}
