% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_data.R
\name{get_location_data}
\alias{get_location_data}
\title{Get data for a location}
\usage{
get_location_data(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data = NULL,
  url = NULL,
  path = NULL,
  package = NULL,
  filetype,
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = NULL,
  crs = NULL,
  ...
)
}
\arguments{
\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Units are passed to
\code{units::set_units()} and then converted to units matching GDAL; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{url}{url for FeatureServer or MapServer layer to pass to
get_area_esri_data. url can be provided to data parameter}

\item{path}{path to spatial data file supported by \code{\link[sf]{read_sf}}}

\item{package}{package name.}

\item{filetype}{file type supported by \code{\link[sf]{read_sf}}. The file
type must be provided for extdata and cached data.}

\item{fn}{Function to apply to data before returning.}

\item{crop}{If TRUE, data cropped to location or bounding box
\code{\link[sf]{st_crop}} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp}
parameters provided. Default TRUE.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default FALSE.}

\item{from_crs}{coordinate reference system of the data.}

\item{crs}{coordinate reference system to return}

\item{...}{additional parameters passed to read_sf_path, read_sf_url, or read_sf_package}
}
\description{
Returns data for a selected area or areas with an optional buffer. If both
crop and trim are FALSE, the function uses \code{\link[sf]{st_intersects}} to
filter data to include the full geometry of anything that overlaps with the
area or bbox (if the area is not provided).
}
