% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_data.R
\name{get_osm_data}
\alias{get_osm_data}
\title{Get Open Street Map features for location}
\usage{
get_osm_data(
  location = NULL,
  key,
  value = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = NULL,
  geometry = NULL,
  osmdata = FALSE
)
}
\arguments{
\item{location}{sf, sfc, or bbox object}

\item{key}{feature key for overpass API query.}

\item{value}{value of feature key; can be negated with an initial exclamation
mark, \code{value = "!this"}, and can also be a vector, \code{value = c ("this", "that")}. If \code{value = "all"} or if \code{key = "building"} the values passed to
osmdata are from \code{osmdata::available_tags(key)}.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{coordinate reference system for output data; if NULL, the data
remains in the OSM CRS 4326. Default: NULL.}

\item{geometry}{geometry to output "polygons", "points", "lines",
"multilines", or "multipolygons"; if NULL return a sf object with mixed
geometry (if multiple geometry types are present). Default NULL}

\item{osmdata}{If TRUE return a osmdata class object that includes the
overpass API call, metadata including timestamp and version numbers, and
all available geometry types; defaults to FALSE.}
}
\description{
Wraps \code{osmdata} functions.
}
