% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_ext.R
\name{get_paper}
\alias{get_paper}
\title{Get paper sizes}
\usage{
get_paper(
  paper = "letter",
  orientation = "portrait",
  standard = NULL,
  series = NULL,
  size = NULL,
  width = NULL,
  height = NULL,
  units = NULL
)
}
\arguments{
\item{paper}{Paper, Default: 'letter'}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'}

\item{standard}{Size standard, "ANSI" or "ISO"}

\item{series}{Size series (e.g. A), Default: \code{NULL}}

\item{size}{Size number (only used for "ISO" and "JIS" series). Standard,
series, and size may all be required to return a single paper using these
parameters.}

\item{width}{Width in units, Default: \code{NULL}}

\item{height}{Height in units, Default: \code{NULL}}

\item{units}{Paper size units, either "in" or "mm"; defaults to \code{NULL} (using
"in" if width or height are provided).}
}
\value{
Data frame with one or more paper sizes.
}
\description{
Use the "paper" parameter (matching name from \link{paper_sizes}), standard
(optionally including both series and size) parameter, or width, height and
units. May return multiple paper sizes.
}
\examples{
\dontrun{
if (interactive()) {
  get_paper("letter")

  get_paper(paper = NULL, standard = "ISO", series = "A", size = 4)

  get_paper(width = 11, height = 17)
}
}
}
