% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_standard_scale}
\alias{get_standard_scale}
\title{Get standard scales and convert to scale distances}
\usage{
get_standard_scale(
  standard = NULL,
  series = NULL,
  scale = NULL,
  paper = NULL,
  orientation = NULL,
  convert = FALSE
)
}
\description{
Get standard scales and convert to scale distances
}
