% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dist_units.R
\name{is_dist_units}
\alias{is_dist_units}
\alias{diff_dist}
\alias{is_diff_dist}
\alias{is_same_dist}
\alias{is_longer}
\alias{is_shorter}
\alias{get_dist_units}
\alias{as_dist_units}
\title{General utility functions for working with distance units objects}
\usage{
is_dist_units(x)

is_diff_dist(x, y, units = NULL)

is_same_dist(x, y, dist = NULL, diff = FALSE, ...)

is_longer(x, y)

is_shorter(x, y)

get_dist_units(x, null.ok = TRUE)

as_dist_units(x, units = NULL, null.ok = FALSE)
}
\arguments{
\item{x, y}{objects to check}

\item{units}{For \link{is_diff_dist}, if x and y are both not units objects, use
units; default to \code{NULL}.}

\item{dist}{type of distance to compare if x and y are sf, sfc, or bbox
objects; "diagdist", "xdist", "ydist". defaults to NULL.}

\item{diff}{If TRUE, return results from \link{is_diff_dist}; if FALSE, return
logical indicator; defaults to FALSE}

\item{...}{Additional parameters passed to all.equal}

\item{null.ok}{If null.ok is \code{TRUE}, allow x to return a \code{NULL} value; if
\code{FALSE}, error on \code{NULL} values.}
}
\description{
\itemize{
\item \link{is_dist_units}: Is x a distance unit object?
\item \link{is_diff_dist}: What is the difference between x and y distance?
\item \link{is_same_dist}: Is x the same distance as y?
\item \link{is_shorter}, is_longer: Is x shorter or longer than y?
}
}
\details{
There are two additional functions that support these utility functions:
\itemize{
\item \link{get_dist_units}: Get the distance units from x (if x is a sf or units
objects or a character string from \link{dist_unit_options})
\item \link{as_dist_units}: Convert x to units using \link[units:units]{units::as_units}
}
}
\seealso{
Other dist: 
\code{\link{convert_dist_scale}()},
\code{\link{convert_dist_units}()},
\code{\link{sf_bbox_misc}},
\code{\link{st_dist}()}
}
\concept{dist}
