% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_sf}
\alias{is_sf}
\alias{is_sfg}
\alias{is_sfc}
\alias{is_bbox}
\alias{is_sf_list}
\alias{is_raster}
\alias{is_sp}
\alias{is_same_crs}
\alias{is_same_area}
\alias{is_sf_or_what}
\title{What is the class or spatial attributes of this feature?}
\usage{
is_sf(x, ext = FALSE, null.ok = FALSE)

is_sfg(x, null.ok = FALSE)

is_sfc(x, null.ok = FALSE)

is_bbox(x, null.ok = FALSE)

is_sf_list(x, named = FALSE, ext = FALSE, null.ok = FALSE)

is_raster(x, null.ok = FALSE)

is_sp(x, null.ok = FALSE)

is_same_crs(x, y)

is_same_area(x, y, union = FALSE, diff = FALSE)

is_sf_or_what(x = NULL, return = NULL, us = FALSE, null.ok = TRUE)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by \link{is_sf})}

\item{null.ok}{If \code{TRUE} and x is \code{NULL}, return \code{TRUE}; defaults to \code{FALSE}.}

\item{y}{An sf object or a character or numeric object supported by
\link[sf:st_crs]{sf::st_crs} that can be compared to x. (used by \link{is_same_crs})}

\item{union}{If \code{TRUE}, use \link[sf:geos_combine]{sf::st_union} on both x and y before comparing
area; defaults to \code{FALSE}.}

\item{diff}{For \link{is_same_dist} and \link{is_same_area}, If \code{TRUE}, return the difference
between the x and y objects. If \code{FALSE}, return \code{TRUE} or \code{FALSE} as expected;
defaults to \code{FALSE}.}

\item{dist}{For \link{is_same_dist}, bounding box dimension to compare distances for "diagdist",
"xdist", or "ydist"; does not currently support numeric distances; defaults
to \code{NULL}.}

\item{is_named}{If \code{TRUE}, check if sf list is named; defaults \code{FALSE}.}
}
\description{
What is the class or spatial attributes of this feature?
}
\details{
\itemize{
\item \link{is_sf}: is x a \code{sf} class object?
\item \link{is_sfc}: is x is a \code{sfc} class object?
\item \link{is_bbox}: is x is a \code{bbox} class object?
\item \link{is_sf_list}: is x is a list of \code{sf} class objects (with or without names)?
\item \link{is_raster}: is x a \code{Raster} class object?
\item \link{is_sp}: is x a \code{Spatial} class object of any type?
\item \link{is_same_crs}: do x and y have the same coordinate reference system?
\item \link{is_same_dist}: do the bbox for x and bbox for y have the same x, y, or diagonal distance?
\item \link{is_same_area}: do x and y have the same area?
}
}
