% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_data.R
\name{layer_location_data}
\alias{layer_location_data}
\title{Layer location data into a ggplot2 map}
\usage{
layer_location_data(
  data = NULL,
  label,
  geom = "sf",
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  package = NULL,
  filetype,
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = NULL,
  crs = NULL,
  ...
)
}
\arguments{
\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{label}{label for area (appended to data as a prefix if data is a
string)}

\item{geom}{ggplot2 geom to use, Default: 'sf'. Options include "sf"
(geom_sf), "icon" (geom_sf_icon), "text" (geom_sf_text), "label"
(geom_sf_label)}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{package}{package name.}

\item{filetype}{file type supported by \code{\link[sf]{read_sf}}. The file
type must be provided for extdata and cached data.}

\item{fn}{Function to apply to data before returning.}

\item{crop}{If TRUE, data cropped to location or bounding box
\code{\link[sf]{st_crop}} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp}
parameters provided. Default TRUE.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default FALSE.}

\item{from_crs}{coordinate reference system of the data.}

\item{crs}{coordinate reference system to return}

\item{...}{Parameters passed to selected geom}
}
\value{
ggplot2 geom
}
\description{
Helper function to make a ggplot2 layer from data returned by
\code{get_location_data}
}
\seealso{
\code{\link[ggplot2]{CoordSf}}
}
