% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_neatline.R
\name{layer_neatline}
\alias{layer_neatline}
\alias{set_map_limits}
\title{Set map limits to sf or bbox object with optional adjustments and panel
border}
\usage{
layer_neatline(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  crs = NULL,
  color = "black",
  size = 1,
  linetype = "solid",
  expand = FALSE,
  hide_grid = TRUE,
  label_axes = "----",
  ...
)
}
\arguments{
\item{data}{\code{sf} or \code{bbox} class object}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Buffer units; defaults to meter.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system to use for
\code{\link[ggplot2]{coord_sf}}.}

\item{color}{Color of panel border, Default: 'black'}

\item{size}{Size of panel border, Default: 1}

\item{linetype}{Line type of panel border, Default: 'solid'}

\item{expand}{Default \code{FALSE.} If \code{TRUE}, the function adds
\code{\link[ggplot2]{scale_y_continuous}} and
\code{\link[ggplot2]{scale_x_continuous}} to expand the map extent to
provided parameters.}

\item{hide_grid}{If \code{TRUE}, hide major grid lines. Default: \code{TRUE}}

\item{label_axes}{A description of which axes to label passed to
\code{\link[ggplot2]{coord_sf}}; defaults to '----' which hides axes
labels.}

\item{...}{Additional parameters to pass to \code{\link[ggplot2]{coord_sf}}.}
}
\value{
\code{\link[ggplot2]{coord_sf}} function with xlim and ylim parameters
}
\description{
Set limits for a map to the bounding box of an x using
\code{\link[ggplot2]{coord_sf}}. Optionally, adjust the x size by applying a
buffer and/or adjust the aspect ratio of the limiting bounding box to match a
set aspect ratio.
}
\examples{
\dontrun{
if (interactive()) {
  library(ggplot2)

  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
  ggplot() +
    geom_sf(data = nc) +
    layer_neatline(data = nc[1, ], asp = 1, color = "red")
}
}
}
\seealso{
\code{\link[ggplot2]{CoordSf}},\code{\link[ggplot2]{scale_continuous}}
}
