% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{mapview_ext}
\alias{mapview_ext}
\alias{mapview_col}
\alias{mapview_yn}
\alias{mapview_exif}
\alias{mapview_flickr}
\alias{make_img_leafpop}
\title{Use mapview to interactively explore spatial data}
\usage{
mapview_col(x, col = NULL, nm = "data", na.rm = TRUE, ...)

mapview_yn(x, col = NULL, no.rm = TRUE, ...)

mapview_exif(path = NULL, filetype = "jpeg", popup = TRUE, ...)

mapview_flickr(
  x = NULL,
  user_id = NULL,
  tags = NULL,
  api_key = NULL,
  img_size = "s",
  num = 25,
  popup = TRUE,
  ...
)

make_img_leafpop(images, popup = TRUE)
}
\arguments{
\item{x}{a \code{Raster*} or \code{Spatial*} or \code{Satellite} or
\code{sf} or \code{stars} object or a list of any combination of those. Furthermore,
  this can also be a \code{data.frame}, a \code{numeric vector} or a
\code{character string} pointing to a tile image folder or file on disk.
  If missing, a blank map will be drawn. A value of NULL will return NULL.}

\item{col}{Column name passed to zcol parameter, Default: \code{NULL}}

\item{nm}{Character vector. If x is a sf list, filter the mapview data to
those named in nm. If \code{NULL}, mapview displays the first item in the sf
list; defaults to "data".}

\item{na.rm}{If TRUE and col is not NULL, filter NA values from the col
before passing to \link[mapview:mapView]{mapview::mapview}}

\item{...}{
  Arguments passed on to \code{\link[mapview:mapView]{mapview::mapview}}
  \describe{
    \item{\code{}}{}
  }}

\item{path}{A path to folder of one or more files with EXIF location
metadata.}

\item{filetype}{The file extension or file type; defaults to \code{NULL}.}

\item{popup}{If \code{TRUE}, add a popup image to a leaflet map; defaults \code{TRUE}.}

\item{user_id}{The NSID of the user with photos to search. If this parameter
is NULL passed then all public photos will be searched.}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags to create an sf object.}

\item{api_key}{Flickr API key. If api_key is `NULL`, the function uses
[getFlickrAPIKey()] to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{img_size}{Defaults to "s" (small). Options ranging from smallest to
largest include "sq", "t", "s", "q", "m", "n", "z", "c", "l", and "o"}

\item{num}{Number of images to display; defaults to 25. 250 is maximum using \code{\link[=get_flickr_photos]{get_flickr_photos()}}.}

\item{images}{A simple feature object with columns for the image path/url, image width, and image height.}
}
\description{
A wrapper for \link[mapview:mapView]{mapview::mapview}. x can be an sf object or sf list. If nm is
NULL and x is an sf list the first item in the list is used.
}
\seealso{
\code{\link[mapview:mapView]{mapview::mapView()}}
}
