% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{mapview_ext}
\alias{mapview_ext}
\alias{mapview_col}
\alias{mapview_exif}
\alias{mapview_flickr}
\alias{make_img_leafpop}
\title{Use mapview to interactively explore spatial data}
\usage{
mapview_col(x, nm = "data", col = NULL, ...)

mapview_exif(path = NULL, filetype = "jpeg", popup = TRUE, ...)

mapview_flickr(
  x = NULL,
  user_id = NULL,
  tags = NULL,
  api_key = NULL,
  img_size = "s",
  num = 25,
  popup = TRUE,
  ...
)

make_img_leafpop(images, popup = TRUE)
}
\arguments{
\item{x}{a \code{Raster*} or \code{Spatial*} or \code{Satellite} or
\code{sf} or \code{stars} object or a list of any combination of those. Furthermore,
  this can also be a \code{data.frame}, a \code{numeric vector} or a
\code{character string} pointing to a tile image folder or file on disk.
  If missing, a blank map will be drawn. A value of NULL will return NULL.}

\item{nm}{Character vector. If x is a sf list, filter the mapview data to
those named in nm. If NULL, mapview displays the first item in the sf list;
defaults to "data".}

\item{col}{Column name passed to zcol parameter, Default: NULL}

\item{...}{
  Arguments passed on to \code{\link[mapview:mapView]{mapview::mapview}}
  \describe{
    \item{\code{}}{}
  }}

\item{path}{path to folder of one or more files with EXIF location metadata}

\item{filetype}{file extension or file type; defaults to "jpg"}

\item{popup}{If \code{TRUE}, add a popup image to a leaflet map; defaults \code{TRUE}.}

\item{user_id}{The NSID of the user who's photo to search. If this parameter
isn't passed then everybody's public photos will be searched.}

\item{tags}{A vector of tags you wish to search for.}

\item{api_key}{Flickr API key. If api_key is `NULL`, the function uses
[getFlickrAPIKey()] to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{img_size}{Defaults to "s" (small). Options ranging from smallest to
largest include "sq", "t", "s", "q", "m", "n", "z", "c", "l", and "o"}
}
\description{
A wrapper for \link[mapview:mapView]{mapview::mapview}
}
\seealso{
\code{\link[mapview]{mapView}}
}
