% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_features.R
\name{number_features}
\alias{number_features}
\alias{sort_features}
\title{Sort and number features}
\usage{
number_features(
  data,
  col = NULL,
  sort = "lon",
  desc = NULL,
  num_style = "arabic",
  suffix = NULL
)

sort_features(data, col = NULL, sort = "lon", desc = NULL)
}
\arguments{
\item{data}{Marker data}

\item{col}{Group column name, Default: \code{NULL}}

\item{sort}{Sort column name, Default: 'lon'}

\item{desc}{If \code{TRUE}, sort descending; set to \code{FALSE} if sort is "lon" and
TRUE is sort is "lat" or any other value, default \code{NULL}}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph", "roman",
"Roman"}

\item{suffix}{Character to appended to "number" column.}
}
\value{
A sf object with a number column ordered by sort.
}
\description{
Used with \code{\link[=layer_numbers]{layer_numbers()}}
}
