% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_exif}
\alias{read_sf_exif}
\title{Read EXIF location data from images to a simple feature object}
\usage{
read_sf_exif(path = NULL, bbox = NULL, filetype = "jpg", sort = "lon", ...)
}
\arguments{
\item{path}{path to folder of one or more files with EXIF location metadata}

\item{bbox}{bounding box to crop sf file (excluding images with location data
outside the bounding box). If bbox is provided the returned data will match
the crs of the bbox.}

\item{filetype}{file extension or file type; defaults to "jpg"}

\item{sort}{variable to sort by. Currently supports "lon" (default), "lat",
or "filename"}

\item{...}{Additional EXIF tags to pass to \code{exiftoolr::exif_read}}
}
\description{
Read EXIF data from folder of images.
}
\seealso{
Other read_write: 
\code{\link{make_filename}()},
\code{\link{read_sf_path}()}
}
\concept{read_write}
