% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_ext}
\alias{read_sf_ext}
\alias{read_sf_pkg}
\alias{read_sf_path}
\alias{read_sf_url}
\alias{read_sf_gsheet}
\alias{read_sf_download}
\title{Read spatial data in a bounding box to a simple feature object}
\usage{
read_sf_ext(..., bbox = NULL)

read_sf_pkg(data, bbox = NULL, package = NULL, filetype = "gpkg", ...)

read_sf_path(path, bbox = NULL, ...)

read_sf_url(url, bbox = NULL, coords = NULL, ...)

read_sf_gsheet(ss, bbox = NULL, coords = c("lon", "lat"), ask = FALSE, ...)

read_sf_download(
  url,
  filename,
  bbox = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  unzip = FALSE,
  ...
)
}
\arguments{
\item{...}{additional parameters passed to multiple functions; see details.}

\item{bbox}{A bounding box object; Default: \code{NULL}. If "bbox" is provided,
read_sf only returns features intersecting the bounding box.}

\item{data}{character; name of data; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only}

\item{package}{character; package name; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only}

\item{filetype}{file type supported by \code{\link[sf:st_read]{sf::read_sf()}}., Default: 'gpkg';
used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only and required only if the data is in the
package cache directory or extdata system files.}

\item{path}{A file path; used by \code{\link[=read_sf_path]{read_sf_path()}} only.}

\item{url}{A url for a spatial data file or for ArcGIS FeatureServer or
MapServer to access with \code{\link[=get_esri_data]{get_esri_data()}}; used by \code{\link[=read_sf_url]{read_sf_url()}} only}

\item{coords}{Character vector with coordinate values; used for
\code{\link[=read_sf_url]{read_sf_url()}} if the "url" is a Google Sheet.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:as_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:sheets_id]{as_sheets_id()}}.}

\item{ask}{If \code{TRUE}, ask for the name of the Google Sheet to read if ss is
not provided to \link{read_sf_gsheet}.}

\item{filename}{File name; if file name is \code{NULL}, name and file type are
both required.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{method}{Method to be used for downloading files.  Current
    download methods are \code{"internal"}, \code{"wininet"} (Windows
    only) \code{"libcurl"}, \code{"wget"} and \code{"curl"}, and there
    is a value \code{"auto"}: see \sQuote{Details} and \sQuote{Note}.

    The method can also be set through the option
    \code{"download.file.method"}: see \code{\link{options}()}.
  }

\item{unzip}{If \code{TRUE}, url must be a zip file that is downloaded, unzipped
into a temporary directory, and then read to a file using the specified
file type.}
}
\description{
An extended version of \code{\link[sf:st_read]{sf::read_sf()}} that support reading spatial
data based on a file path, url (for spatial data or ArcGIS FeatureServer or
MapServer, or a Google Sheet with coordinate columns), or the data name and
associated package. Optionally provide a bounding box to filter data (not
supported for all data types).
}
\details{
Reading data from a package:

\link{read_sf_pkg} looks for three types of package data:
\itemize{
\item Data loaded with the package
\item External data in the \code{extdata} system files folder.
\item Cached data in the cache directory returned by \link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir}
}

Additional ... parameters:

\link{read_sf_pkg} and \link{read_sf_download} both pass additional parameters
to \link{read_sf_path} which supports query, name_col, name, and table. name and
name_col are ignored if a query parameter is provided. If table is not
provided, a expected layer name is created based on the file path.

\link{read_sf_url} pass the where, name_col, and name for any ArcGIS FeatureServer or
MapServer url (passed to \link{get_esri_data}) or sheet if the url is for a Google
Sheet (passed to \link[googlesheets4:range_read]{googlesheets4::read_sheet}), or a query or wkt filter
parameter if the url is some other type (passed to \link[sf:st_read]{sf::read_sf}).

\link{read_sf_ext} is a flexible function that only has bbox as a named parameters
and all other parameters in ... are passed to one of the other overedge
read_sf functions.
}
\seealso{
Other read_write: 
\code{\link{make_filename}()},
\code{\link{read_sf_exif}()}
}
\concept{read_write}
