% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_path}
\alias{read_sf_path}
\alias{read_sf_url}
\alias{read_sf_package}
\title{Read spatial data to simple features object extended  bounding box}
\usage{
read_sf_path(path, bbox = NULL, ...)

read_sf_url(url, bbox = NULL, ...)

read_sf_package(data, bbox = NULL, package, filetype = "gpkg", ...)
}
\arguments{
\item{path}{file path}

\item{bbox}{bbox object; Default: NULL. If bbox is provided, read_sf only
returns features intersecting the bounding box.}

\item{...}{additional parameters passed to \code{\link[sf]{read_sf}}}

\item{url}{url for spatial data file or for ArcGIS FeatureServer or MapServer to access with get_esri_data()}

\item{data}{character; name of data.}

\item{package}{character; package name}

\item{filetype}{file type supported by sf::sf::st_read(), Default: 'gpkg'. Required if the data is in the cache directory or extdata system files.}
}
\value{
sf object
}
\description{
An extended version of \code{\link[sf]{read_sf}} that support reading spatial
data based on a file path, url (for spatial data or ArcGIS FeatureServer or
MapServer), or the data name and associated package. Optionally provide a bounding box
to filter data.
}
\details{
This function looks for three types of package data:
= Data loaded with the package
\itemize{
\item External data in the \code{extdata} system files folder.
\item Cached data in the cache directory returned by \code{\link[rappdirs]{user_cache_dir}}
}
}
