% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_ext}
\alias{read_sf_ext}
\alias{read_sf_path}
\alias{read_sf_url}
\alias{read_sf_pkg}
\alias{read_sf_download}
\title{Read spatial data in a bounding box to a simple feature object}
\usage{
read_sf_path(path, bbox = NULL, ...)

read_sf_url(url, bbox = NULL, coords = NULL, ...)

read_sf_pkg(data, bbox = NULL, package, filetype = "gpkg", ...)

read_sf_download(
  url,
  bbox = NULL,
  filename = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  ...
)
}
\arguments{
\item{path}{A file path; used by \code{\link[=read_sf_path]{read_sf_path()}} only}

\item{bbox}{A bounding box object; Default: \code{NULL}. If "bbox" is provided,
read_sf only returns features intersecting the bounding box.}

\item{...}{additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}. May include query
parameter.}

\item{url}{A url for a spatial data file or for ArcGIS FeatureServer or
MapServer to access with \code{\link[=get_esri_data]{get_esri_data()}}; used by \code{\link[=read_sf_url]{read_sf_url()}} only}

\item{coords}{Character vector with coordinate values; used for
\code{\link[=read_sf_url]{read_sf_url()}} if the "url" is a Google Sheet}

\item{data}{character; name of data; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only}

\item{package}{character; package name; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only}

\item{filetype}{file type supported by \code{\link[sf:st_read]{sf::read_sf()}}., Default: 'gpkg';
used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only and required only if the data is in the
package cache directory or extdata system files.}

\item{filename}{File name; if file name is \code{NULL}, name and file type are
both required.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{method}{Method to be used for downloading files.  Current
    download methods are \code{"internal"}, \code{"wininet"} (Windows
    only) \code{"libcurl"}, \code{"wget"} and \code{"curl"}, and there
    is a value \code{"auto"}: see \sQuote{Details} and \sQuote{Note}.

    The method can also be set through the option
    \code{"download.file.method"}: see \code{\link{options}()}.
  }
}
\description{
An extended version of \code{\link[sf:st_read]{sf::read_sf()}} that support reading spatial
data based on a file path, url (for spatial data or ArcGIS FeatureServer or
MapServer, or a Google Sheet with coordinate columns), or the data name and
associated package. Optionally provide a bounding box to filter data (not
supported for all data types).
}
\details{
Reading data from a package with \code{read_sf_pkg}:

\code{\link[=read_sf_pkg]{read_sf_pkg()}} looks for three types of package data:

= Data loaded with the package
\itemize{
\item External data in the \code{extdata} system files folder.
\item Cached data in the cache directory returned by \link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir}
}
}
\seealso{
Other read_write: 
\code{\link{make_filename}()},
\code{\link{read_sf_exif}()}
}
\concept{read_write}
