% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_buffer_ext.R
\name{st_buffer_ext}
\alias{st_buffer_ext}
\title{Get buffered sf or bbox object}
\usage{
st_buffer_ext(x, dist = NULL, diag_ratio = NULL, unit = "metre")
}
\arguments{
\item{x}{sf or bbox object.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{unit}{Units passed to \code{units::set_units()}.
Default "metre" Must match units of the CRS for the provided object.}
}
\description{
Return an sf object with a buffer based on \code{dist} or a proportion of the
diagonal distance defined by \code{diag_ratio}.
}
