% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_erase}
\alias{st_erase}
\title{Erase an area of a simple feature object}
\usage{
st_erase(x, y, flip = FALSE, union = TRUE)
}
\arguments{
\item{x}{sf object}

\item{y}{sf object}

\item{flip}{If TRUE, use st_intersection; if FALSE use st_difference,
Default: FALSE}

\item{union}{If TRUE, use st_combine and st_union on y before applying
difference/intersection, Default: FALSE}
}
\description{
An extended version of \code{sf::st_difference} that uses the flip parameter to
optionally call \code{sf::st_intersection} and combines and unions the second
object by default.
}
