% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform_ext.R
\name{st_transform_ext}
\alias{st_transform_ext}
\title{Transform or convert coordinates of a simple feature or bounding box object}
\usage{
st_transform_ext(x, crs = NULL, class = NULL)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object, or a character or numeric object
supported by \code{\link[sf]{st_crs}}}

\item{crs}{A coordinate reference system to convert x to or another  \code{sf},
\code{sfc}, or \code{bbox} object that is used to provide crs.}

\item{class}{Class of object to return (sf or bbox).}
}
\value{
An \code{sf}, \code{sfc}, or \code{bbox} object transformed to a new coordinate
reference system.
}
\description{
Take a \code{sf}, \code{sfc}, or \code{bbox} object and transform to coordinate reference
system to match provided \code{crs} or the  CRS for the provided \code{y} object.
}
\seealso{
\code{\link[sf]{st_transform}},\code{\link[sf]{st_crs}}
}
