% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_text.R
\name{theme_text}
\alias{theme_text}
\title{Set, update, or replace text family for ggplot theme and geoms}
\usage{
theme_text(
  font_family = "Helvetica",
  color = "black",
  method = NULL,
  geom = TRUE
)
}
\arguments{
\item{font_family}{Font family, Default: 'Helvetica' If NULL, font_family is
pulled from current set theme which is helpful for resetting all text
families to the theme.}

\item{method}{ggplot2 function to use in changing text "set", "update", or
"replace", Default: 'set'}

\item{geom}{If TRUE, update text family for geom_text, geom_sf_text,
geom_label, and geom_sf_label to match \code{font_family}. If FALSE, make no changes to the theme. Default: TRUE}
}
\description{
Set, update, or replace text family for ggplot theme and geoms
}
\seealso{
\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{theme_get}},\code{\link[ggplot2]{update_geom_defaults}}
}
