% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{write_sf_ext}
\alias{write_sf_ext}
\alias{write_sf_cache}
\title{Write or cache simple feature object to a file}
\usage{
write_sf_ext(
  data,
  name = NULL,
  label = NULL,
  prefix = NULL,
  postfix = NULL,
  filename = NULL,
  filetype = "geojson",
  path = NULL,
  cache = FALSE,
  overwrite = FALSE
)

write_sf_cache(
  data = NULL,
  data_dir = NULL,
  overwrite = FALSE,
  name = NULL,
  label = NULL,
  filetype = NULL,
  filename = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{data}{\code{sf} data to cache.}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor]{make_clean_case}}, e.g. "Residential zoning map"
becomes "residential_zoning_map"}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor]{make_clean_case}}. The label is designed to identify
the area or other shared characteristics across multiple data files, maps,
or plots.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{filename}{File name to use for cached file. Defaults to name of data.
If the data is an \code{sf} object make sure to include the file type, e.g.
"data.gpkg", supported by \code{sf::write_sf()}. All other data is written to
rda with \code{readr::write_rds()}.}

\item{filetype}{File type to write and cache, Default: 'geojson'.}

\item{path}{Path to file or data directory.}

\item{cache}{If \code{TRUE}, write \code{sf} object to file in cache directory; defaults to \code{FALSE}.}

\item{overwrite}{Logical. Default \code{FALSE}. If \code{TRUE}, overwrite any existing
cached files that use the same file name.}

\item{data_dir}{cache data directory, defaults to
\code{\link[rappdirs]{user_cache_dir}} when data_dir is NULL}
}
\description{
write_sf_ext and write_sf_cache wrap the sf::write_sf() function to provide
some additional options including consistent file naming with
\code{\link[=make_filename]{make_filename()}} and additional options including:

Cache data to \code{rappdirs::user_cache_dir("overedge")} or other data directory.
}
\details{
\itemize{
\item Checks that directory exists before attempting to write the file.
\item If filetype is CSV or the filename ends in ".csv" the file is converted to
a dataframe using \code{\link[=df_to_sf]{df_to_sf()}}
\item If cache is \code{TRUE} use write_sf_cache to cache file as well as writing a
copy to the path provided.
}
}
\seealso{
\code{\link[sf]{st_write}}
}
