% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{write_sf_ext}
\alias{write_sf_ext}
\title{Save simple features to files, cache features, or load features to the environment}
\usage{
write_sf_ext(
  data,
  filename = NULL,
  filetype = "geojson",
  data_dir = NULL,
  save = FALSE,
  load = TRUE,
  cache = FALSE
)
}
\arguments{
\item{filetype}{File type to write and cache, Default: 'geojson'}

\item{data_dir}{Data directory to write files, Default: NULL}

\item{save}{If TRUE, write the ; defaults to TRUE}

\item{load}{If TRUE, load sf objects to global environment; defaults to FALSE}

\item{cache}{If TRUE, write sf object(s) to file in cache directory; defaults to FALSE}

\item{x}{sf object or a list of sf objects}
}
\description{
This function wraps \code{sf::write_sf()} and \code{list2env()} to facilitate both
saving simple feature objects and named lists of simple feature objects.
}
\seealso{
\code{\link[sf]{st_write}}
}
