% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{write_sf_ext}
\alias{write_sf_ext}
\alias{write_sf_cache}
\title{Write or cache a simple feature object to a file}
\usage{
write_sf_ext(
  data,
  name = NULL,
  label = NULL,
  prefix = NULL,
  postfix = NULL,
  filename = NULL,
  filetype = "geojson",
  path = NULL,
  cache = FALSE,
  overwrite = FALSE
)

write_sf_cache(
  data,
  data_dir = NULL,
  overwrite = FALSE,
  name = NULL,
  label = NULL,
  filetype = NULL,
  filename = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{data}{\code{sf} object to write.}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor]{make_clean_case}}, e.g. "Residential zoning map"
becomes "residential_zoning_map"}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor]{make_clean_case}}. The label is designed to identify
the area or other shared characteristics across multiple data files, maps,
or plots.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{filename}{File name to use. If filename is provided and the data is an
\code{sf} object make sure to include the file type, e.g. "data.gpkg" or
"data.csv". Objects that are not simple features are written to RDS with
\code{readr::write_rds()}.}

\item{filetype}{File type to write and cache, Default: 'geojson' for
\code{write_sf_ext()}}

\item{path}{Path to file or data directory.}

\item{cache}{If \code{TRUE}, write \code{sf} object to file in cache directory;
defaults to \code{FALSE}.}

\item{overwrite}{Logical. Default \code{FALSE}. If \code{TRUE}, overwrite any existing
cached files that use the same file name.}

\item{data_dir}{cache data directory, defaults to
\code{\link[rappdirs]{user_cache_dir}} when data_dir is NULL. (only used
for write_sf_cache; default is used when cache = TRUE for write_sf_ext)}
}
\description{
The write_sf_ext and write_sf_cache helper functions wrap the
\code{\link[sf]{write_sf}} function to provide some additional options
including consistent file naming with \code{\link[=make_filename]{make_filename()}} and features
including:
\itemize{
\item If the data is not an sf object, optionally save as an RDS file.
\item If filetype is "csv" or the filename ends in ".csv" the file is
automatically converted to a dataframe using \code{\link[=df_to_sf]{df_to_sf()}}; if filetype is
"gsheet" the file is converted and turned into a new Google SHeet document
(if a Google account is authorized with the googlesheets4 package).
\item If cache is \code{TRUE} use write_sf_cache to cache file after writing a copy to
the path provided.
}
}
\seealso{
\code{\link[sf]{st_write}}
}
